% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_to_granges.R
\name{bed_to_granges}
\alias{bed_to_granges}
\title{Read a bed file as a Granges object}
\usage{
bed_to_granges(bed_file)
}
\arguments{
\item{bed_file}{Location of your BED-file}
}
\value{
GR A GRanges object
}
\description{
Read a bed file as a Granges object
}
\examples{
bed_to_granges("my_file.bed")
}

