% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_to_granges_list.R
\name{bed_to_granges_list}
\alias{bed_to_granges_list}
\title{Read a bed file as a Granges object}
\usage{
bed_to_granges_list(bed_file_list)
}
\arguments{
\item{bed_file_list}{List of bed files}
}
\value{
granges_list List of Granges objects
}
\description{
Read a bed file as a Granges object
}

