% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lesion_segretation.R
\name{calculate_lesion_segregation_gr}
\alias{calculate_lesion_segregation_gr}
\title{Calculate the amount of lesion segregation for a singe GRanges object.}
\usage{
calculate_lesion_segregation_gr(
  gr,
  sample_name = "sample",
  split_by_type = FALSE,
  ref_genome = NA
)
}
\arguments{
\item{gr}{GRanges object}

\item{sample_name}{The name of the sample}

\item{split_by_type}{Boolean describing whether the lesion 
segregation should be calculated for all SNVs together or per 96 substitution context.}

\item{ref_genome}{A string matching the name of a BSgenome library
corresponding to the reference genome.
Only needed when split_by_type is TRUE}
}
\value{
A tibble containing the amount of lesions segregation for a single sample
}
\description{
Calculate the amount of lesion segregation for a singe GRanges object.
}
