% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lesion_segretation.R
\name{calculate_strand_switches}
\alias{calculate_strand_switches}
\title{Calculate the total number of variants and strand switches.}
\usage{
calculate_strand_switches(tb)
}
\arguments{
\item{tb}{A tibble with strand information}
}
\value{
A list containing the total number of variants and the number of strand switches
}
\description{
Calculate the total number of variants and strand switches.
}
