% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_mutation_type.R
\name{check_mutation_type}
\alias{check_mutation_type}
\title{Check the mutation type argument}
\usage{
check_mutation_type(type)
}
\arguments{
\item{type}{(Optional) Character vector stating the mutation types to be processed.
If no argument is given, then the function will return 'snv'}
}
\value{
Vector of all possible mutation types from 'type'
}
\description{
A function to check if the mutation type argument contains real 
mutation types which can be processed by this package
}
\examples{
## Check the combination of mutation types
type = c("snv","dbs")
check_mutation_type(type)

}
