% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_remove_muttypes.R
\name{check_no_indels}
\alias{check_no_indels}
\title{Checks that there are no SNVs/MNVs.}
\usage{
check_no_indels(grl)
}
\arguments{
\item{grl}{GRanges/GrangesList object}
}
\value{
Invisibly returns its input when it succeeds.
}
\description{
Checks that there are no SNVs/MNVs.
}
\details{
This function checks for a GRanges/GRangesList object, that there are no Indels.
It trows an error when this is not the case.
}
