% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_remove_muttypes.R
\name{check_no_multi_alts}
\alias{check_no_multi_alts}
\title{Checks that there are no variants with multiple alt alleles.}
\usage{
check_no_multi_alts(grl)
}
\arguments{
\item{grl}{GRanges/GRangesList object}
}
\value{
Invisibly returns its input when it succeeds.
}
\description{
Checks that there are no variants with multiple alt alleles.
}
\details{
This function checks for a GRanges/GRangesList object, that there are no variants with multiple alternative alleles.
It works by checking that the number of variants is equal to the number of alternative alleles.
It trows an error when this is not the case.
}
\examples{

}
