% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_remove_muttypes.R
\name{check_no_snvs_gr}
\alias{check_no_snvs_gr}
\title{Checks that there are no SNVs/MNVs.}
\usage{
check_no_snvs_gr(gr)
}
\arguments{
\item{gr}{GRanges object}
}
\value{
Invisibly returns its input when it succeeds.
}
\description{
Checks that there are no SNVs/MNVs.
}
\details{
This function checks for a single GRanges object, that there are no SNVs/MNVs.
It trows an error when this is not the case.
}
\examples{

}
