% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmic_indel_context.R
\name{cosmic_indel_context}
\alias{cosmic_indel_context}
\title{Retrieve context of indel}
\usage{
cosmic_indel_context(indel)
}
\arguments{
\item{indel}{A data frame with information about a indel. Columns 
needed in data.frame are: \cr
\itemize{
\item{indel_len:} {  Length of an indel sequence}
\item{indel_seq:} {  Sequence of an indel}
\item{indel_type:} {  'del'(etion) or 'ins'(erstion)}
\item{repeats:} {  Number of repeats of an indel}
\item{bimh:} {  Number of bases in microhomology}
}}
}
\value{
Character vector with the context of the indel
}
\description{
A function to extract the indel context from a data.frame, 
specified by the context according to the COSMIC database
}
\examples{
## Get a data.frame with the columns needed. Is also part of the
## function extract_indels
indel = data.frame("chrom" = "chr1",
                   "pos" = 3249117,
                   "ref" = "AT",
                   "alt" = "A",
                   "indel_len" = 1,
                   "indel_type" = "del",
                   "indel_seq" = "T",
                   "repeats" = 4,
                   "bimh" = 1)

context = cosmic_indel_context(indel)

}
\seealso{
\code{\link{extract_indels}}
}
