% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs_count_contexts.R
\name{count_dbs_contexts_gr}
\alias{count_dbs_contexts_gr}
\title{Count dbs contexts from a single GRanges object.}
\usage{
count_dbs_contexts_gr(gr, categories)
}
\arguments{
\item{gr}{GRanges object containing Indel mutations in which the context was added with set_dbs_context.}

\item{categories}{A tibble containing all possible indel context categories}
}
\value{
A single column tibble containing the number of dbs per COSMIC context.
}
\description{
Count dbs contexts from a single GRanges object.
}
\details{
Counts the number of DBS per COSMIC context from a GRanges object containing DBS mutations.
The function is called by count_dbs_contexts
}
\seealso{
\code{\link{count_dbs_contexts}}, \code{\link{set_dbs_context}}

Other DBS: 
\code{\link{count_dbs_contexts}()},
\code{\link{plot_dbs_contexts}()},
\code{\link{plot_main_dbs_contexts}()},
\code{\link{set_context_dbs_gr}()},
\code{\link{set_context_dbs}()}
}
\concept{DBS}
