% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indel_count_contexts.R
\name{count_indel_contexts}
\alias{count_indel_contexts}
\title{Count indel contexts}
\usage{
count_indel_contexts(grl)
}
\arguments{
\item{grl}{GRanges or GRangesList object containing Indel mutations in which the context was added with get_indel_context.}
}
\value{
A tibble containing the number of indels per COSMIC context per gr.
}
\description{
Count indel contexts
}
\details{
Counts the number of indels per COSMIC context from a GRanges or GRangesList object containing Indel mutations.
This function applies the count_indel_contexts_gr function to each gr in its input.
It then combines the results in a single tibble and returns this.
}
\examples{

}
\seealso{
\code{\link{count_indel_contexts_gr}}, \code{\link{get_indel_context}}

Other Indels: 
\code{\link{count_indel_contexts_gr}()},
\code{\link{get_1bp_dels}()},
\code{\link{get_1bp_ins}()},
\code{\link{get_big_dels}()},
\code{\link{get_big_ins}()},
\code{\link{get_extended_sequence}()},
\code{\link{get_indel_context_gr}()},
\code{\link{get_indel_context}()},
\code{\link{plot_indel_contexts}()},
\code{\link{plot_main_indel_contexts}()}
}
\concept{Indels}
