% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indel_count_contexts.R
\name{count_indel_contexts_gr}
\alias{count_indel_contexts_gr}
\title{Count indel contexts from a single gr.}
\usage{
count_indel_contexts_gr(gr, categories)
}
\arguments{
\item{gr}{GRanges object containing Indel mutations in which the context was added with get_indel_context.}

\item{categories}{A tibble containing all possible indel context categories}
}
\value{
A single column tibble containing the number of indels per COSMIC context.
}
\description{
Count indel contexts from a single gr.
}
\details{
Counts the number of indels per COSMIC context from a GRanges object containing Indel mutations.
The function is called by count_indel_contexts_gr
}
\examples{

}
\seealso{
\code{\link{count_indel_contexts}}, \code{\link{get_indel_context}}

Other Indels: 
\code{\link{count_indel_contexts}()},
\code{\link{get_1bp_dels}()},
\code{\link{get_1bp_ins}()},
\code{\link{get_big_dels}()},
\code{\link{get_big_ins}()},
\code{\link{get_extended_sequence}()},
\code{\link{get_indel_context_gr}()},
\code{\link{get_indel_context}()},
\code{\link{plot_indel_contexts}()},
\code{\link{plot_main_indel_contexts}()}
}
\concept{Indels}
