% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbs_count_contexts.R
\name{count_mbs_contexts_gr}
\alias{count_mbs_contexts_gr}
\title{Count MBS grouped by length from a single GRanges object.}
\usage{
count_mbs_contexts_gr(gr, categories)
}
\arguments{
\item{gr}{GRanges object containing mbs variants.}

\item{categories}{A tibble containing all possible mbs size categories}
}
\value{
A single column tibble containing the number of MBS per MBS length
}
\description{
Count MBS grouped by length from a single GRanges object.
}
\details{
Counts the number of MBS per MBS length from a GRanges object containing mbs variants.
The function is called by count_mbs_contexts
}
\seealso{
\code{\link{count_mbs_contexts}}

Other MBS: 
\code{\link{count_mbs_contexts}()},
\code{\link{plot_mbs_contexts}()}
}
\concept{MBS}
