% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_type_occurences_old.R, R/count_types_occurences.R
\name{count_type_occurences}
\alias{count_type_occurences}
\title{Count the occurences of each substitution type}
\usage{
count_type_occurences(vcf_list, ref_genome)

count_type_occurences(vcf_list, ref_genome)
}
\arguments{
\item{vcf_list}{A list of CollapsedVCF object}

\item{vcf_list}{A list of CollapsedVCF object}
}
\value{
dataframe

dataframe
}
\description{
Count the occurences of each substitution type

Count the occurences of each substitution type
}

