% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_rank.R
\name{estimate_rank}
\alias{estimate_rank}
\title{Estimate optimal rank for NMF decomposition}
\usage{
estimate_rank(mut_spectrum, rank_range, nrun = 100, outdir)
}
\arguments{
\item{rank_range}{Range of ranks one would like to test}

\item{outdir}{Specify output directory}

\item{vcf_files}{Character vector of vcf_files}
}
\description{
Find optimal number of signatures for NMF decomposition
}

