% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos_sim_matrix.R
\name{explained_by_signatures}
\alias{explained_by_signatures}
\title{This function has been renamed to 'cos_sim_matrix'.}
\usage{
explained_by_signatures(mut_matrix, signatures)
}
\arguments{
\item{mut_matrix}{96 mutation count matrix (dimensions: 96 mutations X n samples)}

\item{signatures}{96 mutation count matrix (dimensions: 96 mutations X m samples)}
}
\value{
Matrix with pairwise cosine similarities
}
\description{
This function has been renamed to 'cos_sim_matrix'.
}
\examples{
# You can download the signatures from the COSMIC website:
# https://cancer.sanger.ac.uk/cosmic/signature
## We copied the file into our package for your convenience.
filename <- system.file("extdata/snv_signatures_probabilities.txt",
  package = "MutationalPatterns"
)
signatures <- read.table(filename, sep = "\t", header = TRUE)
signatures <- as.matrix(signatures[, -c(1, 2)])

## See the 'mut_matrix()' example for how we obtained the mutation matrix:
mut_mat <- readRDS(system.file("states/mut_mat_data.rds",
  package = "MutationalPatterns"
))


## Calculate the cosine similarity between each COSMIC signature and each 96 mutational profile
cos_sim_matrix(mut_mat, signatures)
}
\seealso{
\code{\link{cos_sim_matrix}}
\code{\link{mut_matrix}},
\code{\link{fit_to_signatures}},
\code{\link{plot_cosine_heatmap}}
}
