% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_indels.R
\name{extract_indels}
\alias{extract_indels}
\title{Extract indel signatures}
\usage{
extract_indels(bed, context.database = "cosmic", sample.name = NULL,
  ref.genome = DEFAULT_GENOME, indel.len.cap = 6, n.bases.mh.cap = 5,
  verbose = F, ...)
}
\arguments{
\item{bed}{A dataframe containing the columns: chrom, pos, ref, alt}

\item{context.database}{(Optional) A character naming the context database of the indels. Options are "cosmic" to get the contexts from COSMIC
or "predefined" to get the predefined context from MutationalPatterns.\cr
Default is "cosmic"}

\item{sample.name}{(Optional) If a character is provided, the header for the output matrix will be named to this. If none is
provided, the basename of the vcf file will be used.}

\item{ref.genome}{(Optional) A character naming the BSgenome reference genome. Default is "BSgenome.Hsapiens.UCSC.hg19". If another
reference genome is indicated, it will also need to be installed.}

\item{indel.len.cap}{(Optional) Specifies the max indel sequence length to consider when counting 'repeat' and 'none' contexts.
Counts of longer indels will simply be binned to the counts of contexts at the max indel sequence length.}

\item{n.bases.mh.cap}{(Optional) Specifies the max bases in microhomology to consider when counting repeat and microhomology
contexts. Counts of longer indels will simply be binned to the counts of contexts at the max indel sequence length.}

\item{verbose}{(Optional) Print progress messages?}

\item{...}{Other arguments that can be passed to get_contexts_indel()}
}
\value{
A 1-column mutation count matrix
}
\description{
Get a 1-column matrix containing the absolute indel signature
contributions (i.e. the number of mutations contributing to each mutational signature).\cr
For the context of the indels, there are two options:\cr\cr
Use 'cosmic' for signatures defined by the COSMIC database.\cr
Use 'predefined' for signatures defined as insertions/deletions within repeat regions (ins.rep, del.rep),
insertions/deletions with flanking microhomology (ins.mh, del.mh), and insertions/deletions
which don't fall under the previous 2 categories (ins.none, del.none). Each category is further
stratified by the length of the indel.
}
