% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sample_name.R
\name{extract_sample_name}
\alias{extract_sample_name}
\title{Extract sample name from path to file}
\usage{
extract_sample_name(filename)
}
\arguments{
\item{file_name}{String of path to file}
}
\value{
name String of sample name
}
\description{
Function to extract the sample name from path to file
}

