% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate_muts.R
\name{extrapolate_muts}
\alias{extrapolate_muts}
\title{Extract sample name from path to file}
\usage{
extrapolate_muts(vcf, GR_surveyed, genome_length)
}
\arguments{
\item{vcf}{CollapsedVCF object}

\item{GR_surveyed}{GRanges object of surveyed bed file}
}
\value{
df Dataframe with number of mutations extrapolated to whole genome and other stats
}
\description{
Function to extract the sample name from path to file
}

