% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_nested.R
\docType{data}
\name{facet_nested}
\alias{facet_nested}
\alias{FacetNested}
\title{Plot merged facets}
\usage{
facet_nested(rows = NULL, cols = NULL, scales = "fixed",
  labeller = "label_value")
}
\arguments{
\item{rows}{Variables on the rows}

\item{cols}{Variables on the columns}

\item{scales}{Character string indicating if scale of x or y axis is fixed or free}

\item{labeller}{Character vector with the labels of the facets}
}
\description{
Helper function for ggplot to merge facets
}
\keyword{datasets}
