% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcfs_as_granges.R
\name{filter_seqlevels}
\alias{filter_seqlevels}
\title{Filter a GRanges object based on seqlevels}
\usage{
filter_seqlevels(gr, group, genome)
}
\arguments{
\item{gr}{GRanges object}

\item{group}{Selector for a seqlevel group.  All seqlevels outside
of this group will be removed.  Possible values:
* 'auto' for autosomal chromosomes;
* 'sex' for sex chromosomes;
* 'auto+sex' for autosomal + sex chromosomes (default);
* 'circular' for circular chromosomes;
* 'none' for no filtering, which results in keeping all
  seqlevels from the VCF file.}

\item{genome}{BSgenome object}
}
\value{
A GRanges object
}
\description{
This function filters a GRanges object based on a group of seqnames.
}
