% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_remove_muttypes.R
\name{find_snv}
\alias{find_snv}
\title{Identifies SNVs/MNVs}
\usage{
find_snv(gr)
}
\arguments{
\item{gr}{GRanges object}
}
\value{
A boolean vector. It's TRUE for SNVs/MNVs and FALSE for Indels
}
\description{
Identifies SNVs/MNVs
}
\details{
This function finds SNVs/MNVs for a single GRanges object.
}
