% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_distribution_list.R
\name{genomic_distribution_list}
\alias{genomic_distribution_list}
\title{Find overlaps between mutations and a genomic region}
\usage{
genomic_distribution_list(MutPat_object, region_list)
}
\arguments{
\item{MutPat_object}{A list...}

\item{region_list}{List with GRanges objects containing locations of genomic regions}
}
\description{
Function finds the number of mutations that reside in genomic region and takes surveyed area of genome into account
}

