% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indel_get_context.R
\name{get_big_dels}
\alias{get_big_dels}
\title{Get contexts from larger deletions}
\usage{
get_big_dels(gr, mut_size, ref_genome)
}
\arguments{
\item{gr}{GRanges object containing Indel mutations. 
The mutations should be called similarly to HaplotypeCaller.}

\item{mut_size}{A double vector containing the size of each Indel.}

\item{ref_genome}{BSGenome reference genome object}
}
\value{
A modified version of the input gr. 
All variants that are not a deletion larger than 1bp are removed.
In each gr two columns have been added. 
"muttype" showing the main indel type and "muttype_sub" which shows the subtype. 
The subtype is the number of repeats or the microhomology length.
}
\description{
Get contexts from larger deletions
}
\details{
Determines the COSMIC context for deletions larger than 1bp in a GRanges object containing Indel mutations.
This function is called by get_indel_context_gr.
The function determines if there is microhomology for deletions that are not in repeat regions.
}
\seealso{
\code{\link{get_indel_context_gr}}

Other Indels: 
\code{\link{count_indel_contexts_gr}()},
\code{\link{count_indel_contexts}()},
\code{\link{get_1bp_dels}()},
\code{\link{get_1bp_ins}()},
\code{\link{get_big_ins}()},
\code{\link{get_extended_sequence}()},
\code{\link{get_indel_context_gr}()},
\code{\link{get_indel_context}()},
\code{\link{plot_indel_contexts}()},
\code{\link{plot_main_indel_contexts}()}
}
\concept{Indels}
