% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indel_get_context.R
\name{get_big_ins}
\alias{get_big_ins}
\title{Get contexts from bigger inserions}
\usage{
get_big_ins(gr, mut_size, ref_genome)
}
\arguments{
\item{gr}{GRanges object containing Indel mutations. 
The mutations should be called similarly to HaplotypeCaller.}

\item{mut_size}{A double vector containing the size of each Indel.}

\item{ref_genome}{BSGenome reference genome object}
}
\value{
A modified version of the input gr. 
All variants that are not insertions larger than 1bp are removed.
In each gr two columns have been added. 
"muttype" showing the main indel type and "muttype_sub" which shows the subtype. 
The subtype is the number of repeats.
}
\description{
Get contexts from bigger inserions
}
\details{
Determines the COSMIC context for insertions larger than 1bp in a GRanges object containing Indel mutations.
This function is called by get_indel_context_gr.
}
\seealso{
\code{\link{get_indel_context_gr}}

Other Indels: 
\code{\link{count_indel_contexts_gr}()},
\code{\link{count_indel_contexts}()},
\code{\link{get_1bp_dels}()},
\code{\link{get_1bp_ins}()},
\code{\link{get_big_dels}()},
\code{\link{get_extended_sequence}()},
\code{\link{get_indel_context_gr}()},
\code{\link{get_indel_context}()},
\code{\link{plot_indel_contexts}()},
\code{\link{plot_main_indel_contexts}()}
}
\concept{Indels}
