% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contexts_indel.R
\name{get_contexts_indel}
\alias{get_contexts_indel}
\title{Extract indel sequence, type and length}
\usage{
get_contexts_indel(bed, ref_genome = DEFAULT_GENOME,
  get_other_indel_allele = F, verbose = F)
}
\arguments{
\item{bed}{A dataframe containing the columns: chrom, pos, ref, alt}

\item{ref_genome}{(Optional) A character naming the BSgenome reference genome. Default is
"BSgenome.Hsapiens.UCSC.hg19". If another reference genome is indicated, it will also need to be
installed.}

\item{get_other_indel_allele}{(Optional) Only applies when mode=='indel' For indels, some vcfs only report
the sequence of one allele (REF for deletions and ALT for insertions). If TRUE, the unreported
allele will be retrieved from the genome: a 5' base relative to the indel sequence. This base
will also be added to the indel sequence and the POS will be adjusted accordingly (POS=POS-1).}

\item{verbose}{(Optional) Print progress messages?}
}
\value{
A dataframe in the same structure as a bed file
}
\description{
Extract indel sequence, type and length
}
