% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indel_get_context.R
\name{get_extended_sequence}
\alias{get_extended_sequence}
\title{Retreive the flanking sequence from a GRanges object}
\usage{
get_extended_sequence(gr, flank_dist, ref_genome)
}
\arguments{
\item{gr}{GRanges object containing Indel mutations. 
The mutations should be called similarly to HaplotypeCaller.}

\item{flank_dist}{A numeric vector of length one containing the number of flanking base pairs.}

\item{ref_genome}{BSGenome reference genome object}
}
\value{
A DNAStringSet containing the flaning bases.
}
\description{
Retreive the flanking sequence from a GRanges object
}
\details{
This function retreives the flanking sequence from a GRanges object.
The size of the flanking sequence is supplied as an argument to the function.
This function works by first extending the ranges and then retreiving the sequence.
}
\examples{

}
\seealso{
Other Indels: 
\code{\link{count_indel_contexts_gr}()},
\code{\link{count_indel_contexts}()},
\code{\link{get_1bp_dels}()},
\code{\link{get_1bp_ins}()},
\code{\link{get_big_dels}()},
\code{\link{get_big_ins}()},
\code{\link{get_indel_context_gr}()},
\code{\link{get_indel_context}()},
\code{\link{plot_indel_contexts}()},
\code{\link{plot_main_indel_contexts}()}
}
\concept{Indels}
