% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indel_get_context.R
\name{get_indel_context}
\alias{get_indel_context}
\title{Get indel contexts}
\usage{
get_indel_context(grl, ref_genome)
}
\arguments{
\item{grl}{GRanges or GRangesList object containing Indel mutations. 
The mutations should be called similarly to HaplotypeCaller.}

\item{ref_genome}{BSGenome reference genome object}
}
\value{
A modified version of the input grl. In each gr two columns have been added. 
"muttype" showing the main indel type and "muttype_sub" which shows the subtype. 
The subtype is either the number of repeats or the microhomology length.
}
\description{
Get indel contexts
}
\details{
Determines the COSMIC context from a GRanges or GRangesList object containing Indel mutations.
It applies the get_indel_context_gr function to each gr in the input.
}
\examples{

}
\seealso{
\code{\link{get_indel_context_gr}},

Other Indels: 
\code{\link{count_indel_contexts_gr}()},
\code{\link{count_indel_contexts}()},
\code{\link{get_1bp_dels}()},
\code{\link{get_1bp_ins}()},
\code{\link{get_big_dels}()},
\code{\link{get_big_ins}()},
\code{\link{get_extended_sequence}()},
\code{\link{get_indel_context_gr}()},
\code{\link{plot_indel_contexts}()},
\code{\link{plot_main_indel_contexts}()}
}
\concept{Indels}
