% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indel_get_context.R
\name{get_indel_context_gr}
\alias{get_indel_context_gr}
\title{Get indel contexts from a single gr}
\usage{
get_indel_context_gr(gr, ref_genome)
}
\arguments{
\item{gr}{GRanges object containing Indel mutations. 
The mutations should be called similarly to HaplotypeCaller.}

\item{ref_genome}{BSGenome reference genome object}
}
\value{
A modified version of the input gr. In the gr two columns have been added. 
"muttype" showing the main indel type and "muttype_sub" which shows the subtype. 
The subtype is either the number of repeats or the microhomology length.
}
\description{
Get indel contexts from a single gr
}
\details{
Determines the COSMIC context from a GRanges object containing Indel mutations.
It throws an error if there are any variants with multiple alternative alleles or SNVs.
}
\seealso{
\code{\link{get_indel_context}},

Other Indels: 
\code{\link{count_indel_contexts_gr}()},
\code{\link{count_indel_contexts}()},
\code{\link{get_1bp_dels}()},
\code{\link{get_1bp_ins}()},
\code{\link{get_big_dels}()},
\code{\link{get_big_ins}()},
\code{\link{get_extended_sequence}()},
\code{\link{get_indel_context}()},
\code{\link{plot_indel_contexts}()},
\code{\link{plot_main_indel_contexts}()}
}
\concept{Indels}
