% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mut_type.R
\name{get_mut_type_gr}
\alias{get_mut_type_gr}
\title{Get variants with mut_type from GRanges}
\usage{
get_mut_type_gr(gr, type = c("snv", "indel", "dbs", "mbs"))
}
\arguments{
\item{gr}{GRanges}

\item{type}{The type of variant that will be returned.}
}
\value{
GRanges of the desired mutation type.
}
\description{
Get the variants of a certain mutation type from a GRanges object.
All other variants will be filtered out.
It is assumed that DBS/MBSs are called as separate SNVs.
They are merged into single variants.
The type of variant can be chosen with type.
}
\seealso{
\code{\link{get_mut_type}}, \code{\link{split_mbs_gr}}, \code{\link{read_vcfs_as_granges}}
}
