% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_muts.R
\name{get_muts}
\alias{get_muts}
\title{Retrieve base substitutions from vcf file}
\usage{
get_muts(vcf)
}
\arguments{
\item{vcf}{A CollapsedVCF object}
}
\value{
muts Base substitutions
}
\description{
A function to extract REF and ALT alleles from vcf file
}

