% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_muts_region.R
\name{get_muts_region}
\alias{get_muts_region}
\title{Split a GRanges object based on a list of regions.}
\usage{
get_muts_region(gr, ranges)
}
\arguments{
\item{gr}{GRanges object containing variants}

\item{ranges}{GRanges object containing regions of interest}
}
\value{
GRanges object
}
\description{
A GRanges object containing variants is split based on a list of regions.
This list can be either a GRangesList or a GRanges object.
The result is a GRangesList where each element contains the variants of one sample from one region.
Variant that are not in any of the provided region are put in a list of 'other'.
}
\seealso{
Other genomic_regions: 
\code{\link{lengthen_mut_matrix_single_sample}()},
\code{\link{lengthen_mut_matrix}()},
\code{\link{plot_spectrum_region}()},
\code{\link{split_muts_region_gr}()},
\code{\link{split_muts_region}()}
}
\concept{genomic_regions}
