% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sig_start.R
\name{get_sig_star}
\alias{get_sig_star}
\title{Determine the number of significance stars}
\usage{
get_sig_star(val, cutoffs)
}
\arguments{
\item{val}{Statistical value. Either a p value or fdr.}

\item{cutoffs}{Significance cutoffs for the statistical value.}
}
\value{
A vector of significance stars and empty strings (not significant).
}
\description{
The number of significance stars is determined based on the statistical value
and the significance cutoffs.
}
