% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_class.R
\name{get_sim_tb}
\alias{get_sim_tb}
\alias{get_sim_tb,region_cossim-method}
\title{An S4 generic to get the sim_tb from a region_cossim object.}
\usage{
get_sim_tb(x)

\S4method{get_sim_tb}{region_cossim}(x)
}
\arguments{
\item{x}{A region_cossim object}

\item{region_cossim}{A region_cossim object}
}
\value{
A tibble containing the calculated similarities of the windows.

A tibble containing the calculated similarities of the windows.
}
\description{
An S4 generic to get the sim_tb from a region_cossim object.

An S4 method for the get_sim_tb generic
}
\section{Methods (by class)}{
\itemize{
\item \code{get_sim_tb(region_cossim)}: Get the sim_tb from a region_cossim object.

}}
