% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_strand.R
\name{get_strand}
\alias{get_strand}
\title{Find transcriptional strand of base substitutions in vcf}
\usage{
get_strand(vcf, genes)
}
\arguments{
\item{vcf}{Granges vcf object}

\item{genes}{Granges with definition of gene bodies, should include strand information}
}
\value{
Character vector with transcriptional strand information with length of vcf: "-" for positions outside gene bodies, "U" for untranscribed/sense/coding strand, "T" for transcribed/anti-sense/non-coding strand
}
\description{
For the positions that are within gene bodies it is determined whether the "C" or "T" base 
(since by convention we regard base substitutions as C>X or T>X) are on the same strand as the gene definition.
Base substitions on the same strand as the gene definitions are considered "untranscribed", and on the opposite strand of gene bodies as transcribed, 
since the gene definitions report the coding or sense strand, which is untranscribed.  
No strand information "-" is reported for base substitution that overlap with more than one gene body.
}

