% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lesion_segretation.R
\name{get_strandedness_tb}
\alias{get_strandedness_tb}
\title{Convert a GRanges object with strand info to a tibble}
\usage{
get_strandedness_tb(gr)
}
\arguments{
\item{gr}{A GRanges object where the strands have been set.}
}
\value{
A tibble with strand information
}
\description{
Convert a GRanges object with strand info to a tibble
}
\seealso{
Other Lesion_segregation: 
\code{\link{calculate_lesion_segregation_gr}()},
\code{\link{calculate_lesion_segregation}()},
\code{\link{calculate_strand_switches}()},
\code{\link{calculate_strand_switch}()},
\code{\link{get_strandedness_gr}()},
\code{\link{lesion_get_x_axis_breaks}()},
\code{\link{plot_lesion_segregation_gg}()},
\code{\link{plot_lesion_segregation}()}
}
\concept{Lesion_segregation}
