% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_types.R
\name{get_types}
\alias{get_types}
\title{Retrieve base substitution types from vcf file}
\usage{
get_types(vcf)
}
\arguments{
\item{vcf}{A CollapsedVCF object}
}
\value{
types Base substitution types
}
\description{
A function to extract the base substitutions and translate to the 6 base substitution types
}

