% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indel_mutation_type.R
\name{indel_mutation_type}
\alias{indel_mutation_type}
\title{Set global context, class and color values for indels}
\usage{
indel_mutation_type(indel)
}
\arguments{
\item{indel}{(Optional) List of vectors for context, class and color of indels.
Color vector must have the same length as the class vector, 
because in plotting profiles, each class is represented by one color.\cr\cr
It is also possible to give a character for predefined variables:
\itemize{
  \item{"predefined"} {Represents a indel context of 3 classes per 
  deletion and indel: "repetitive region", "microhomology" and 
  "none" of these two. Indels have lengths 1 to 5+}
  \item{"cosmic"} {Represents the indel context according to the
  COSMIC database}
}
Default is "cosmic"}
}
\value{
The global variables indel_class, indel_class_header,
indel_context and indel_colors will be set with this function. 
Furthermore a character is returned, which state the chosen option
}
\description{
A function to set the global values of the context variable, 
the class variable and the color variable of indels. For indels
there is no single intuitive and naturally constrained set of 
mutation types, so a custom set of mutation types can be used
}
\examples{
## For custom input of predefined context
indel = list("matrix" = mut_matrix,
             "context" =c(
paste0('del.rep.len.', 1:5),
paste0('ins.rep.len.', 1:5),
paste0('del.mh.bimh.', 1:5),
paste0('ins.mh.bimh.', 1:5),
paste0('del.none.len.', 1:5),
paste0('ins.none.len.', 1:5)
),
             "class"=c(
rep('del.rep', 5), rep('ins.rep', 5),
rep('del.mh', 5), rep('ins.mh', 5),
rep('del.none', 5), rep('ins.none', 5)
),
             "colors" = c(
"#F7BF80", "#ED8212", 
"#B5D988", "#31A12C", 
"#E44A39", "#B81C20"
))

indel_mutation_type(indel)

}
