% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_with_region.R
\name{intersect_with_region}
\alias{intersect_with_region}
\title{Find overlap between mutations and a genomic region}
\usage{
intersect_with_region(vcf, surveyed, region)
}
\arguments{
\item{vcf}{CollapsedVCF object with mutations}

\item{surveyed}{GRanges object with regions of the genome that were surveyed}

\item{region}{GRanges object with genomic region(s)}
}
\value{
A Data.frame containing the overlapping mutations for a genomic region.
}
\description{
Find the number of mutations that reside in genomic region and take surveyed area of genome into account
}

