% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthen_mut_matrix.R
\name{lengthen_mut_matrix_single_sample}
\alias{lengthen_mut_matrix_single_sample}
\title{Lengthen mutation matrix for a single sample}
\usage{
lengthen_mut_matrix_single_sample(mut_matrix)
}
\arguments{
\item{mut_matrix}{Mutation matrix}
}
\value{
mut_matrix single sample
}
\description{
This function is called by 'lengthen_mut_matrix()' to lengthen a single sample.
}
\seealso{
\code{\link{lengthen_mut_matrix}}

Other genomic_regions: 
\code{\link{get_muts_region}()},
\code{\link{lengthen_mut_matrix}()},
\code{\link{plot_long_profile}()},
\code{\link{plot_spectrum_region}()},
\code{\link{split_muts_region_gr}()},
\code{\link{split_muts_region}()}
}
\concept{genomic_regions}
