% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lesion_segregation.R
\name{lesion_get_x_axis_breaks}
\alias{lesion_get_x_axis_breaks}
\title{Determine the x axis breaks of a lesion segregation plot}
\usage{
lesion_get_x_axis_breaks(max_coord, per_chrom)
}
\arguments{
\item{max_coord}{Maximum coordinate value of a variant}

\item{per_chrom}{Boolean. Describing whether to create a separate plot per chromosome}
}
\value{
Numeric vector of x_axis_breaks
}
\description{
The maximum x axis value is determined by the variant with the highest coordinate.
Additionally, there are more breaks when the plot is created per chromosome.
}
\seealso{
Other Lesion_segregation: 
\code{\link{calculate_lesion_segregation}()},
\code{\link{plot_lesion_segregation_gg}()},
\code{\link{plot_lesion_segregation}()}
}
\concept{Lesion_segregation}
