% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_matrix_NMF.R
\name{make_matrix_NMF}
\alias{make_matrix_NMF}
\title{Construct trinucleotide changes frequency matrix}
\usage{
make_matrix_NMF(vcf_list, type = "normal")
}
\arguments{
\item{vcf_list}{List of collapsed vcf objects from which one would like to contstruct a count matrix}
}
\value{
df Count data.frame
}
\description{
Construct trinucleotide changes frequency matrix
}

