% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mut_matrix.R
\name{make_mut_matrix}
\alias{make_mut_matrix}
\title{Make 96 trinucleotide count matrix}
\usage{
make_mut_matrix(vcf_list, ref_genome)
}
\arguments{
\item{vcf_list}{List of collapsed vcf objects from which one would like to contstruct a count matrix}
}
\value{
96 mutation count matrix
}
\description{
Make 96 trinucleotide mutation count matrix
}

