% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sample_row_192.R
\name{make_sample_row_192}
\alias{make_sample_row_192}
\title{Make sample row 192 for NMF}
\usage{
make_sample_row_192(vcf)
}
\arguments{
\item{vcf}{Collapsed vcf object}
}
\value{
vector
}
\description{
Make sample row 192 for NMF
}

