% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sample_row_288.R
\name{make_sample_row_288}
\alias{make_sample_row_288}
\title{Make sample row 288}
\usage{
make_sample_row_288(vcf)
}
\arguments{
\item{signatures}{Dataframe containing mutational sigantures}
}
\description{
Make sample row 288
}

