% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sample_row_96.R
\name{make_sample_row_96}
\alias{make_sample_row_96}
\title{Make sample row for NMF input matrix}
\usage{
make_sample_row_96(type_context)
}
\arguments{
\item{type_context}{result from get_type_context function}
}
\value{
matrix
}
\description{
Make sample row for NMF input matrix
}

