% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mut_type.R
\name{merge_muts}
\alias{merge_muts}
\title{Merge variants in a GRanges object.}
\usage{
merge_muts(gr)
}
\arguments{
\item{gr}{GRanges object}
}
\value{
GRanges object with a single variant.
}
\description{
Merge all variants in a GRanges object into a single variant.
This is usefull when a DBS/MNV is called as separate SNVs.
}
\seealso{
\code{\link{split_mbs_gr}}
}
