% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_96_occurences.R
\name{mut_96_occurences}
\alias{mut_96_occurences}
\title{Count 96 trinucleotide mutation occurences}
\usage{
mut_96_occurences(type_context)
}
\arguments{
\item{type_context}{result from get_type_context function}
}
\value{
vector with 96 trinucleotide mutation occurences
}
\description{
Count 96 trinucleotide mutation occurences
}

