% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_96_occurrences.R
\name{mut_96_occurrences}
\alias{mut_96_occurrences}
\title{Count 96 trinucleotide mutation occurrences}
\usage{
mut_96_occurrences(type_context, gr_sizes)
}
\arguments{
\item{type_context}{result from type_context function}

\item{gr_sizes}{A vector indicating the number of variants per GRanges}
}
\value{
Mutation matrix with 96 trinucleotide mutation occurrences
}
\description{
@details
 This function is called by mut_matrix. It calculates the 96 trinucleotide context for all variants
 and then splits these per GRanges (samples). It then calculates how often each 96 trinucleotide context occurs.
}
