% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_type_occurences.R
\name{mut_type_occurences}
\alias{mut_type_occurences}
\title{Count the occurences of each base substitution type}
\usage{
mut_type_occurences(vcf_list, ref_genome)
}
\arguments{
\item{vcf_list}{A list of CollapsedVCF object}
}
\value{
Dataframe with counts of each base substitution type for each sample in vcf_list
}
\description{
Count the occurences of each base substitution type
}

