% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_type.R
\name{mutation_types}
\alias{mutation_types}
\title{This function has been renamed.  Use 'mut_type' instead.}
\usage{
mutation_types(vcf)
}
\arguments{
\item{vcf}{A GRanges object.}
}
\value{
Character vector with base substitution types
}
\description{
This function has been renamed.  Use 'mut_type' instead.
}
\examples{
## See the 'read_vcfs_as_granges()' example for how we obtained the
## following data:
vcfs <- readRDS(system.file("states/read_vcfs_as_granges_output.rds",
                package="MutationalPatterns"))

mut_type(vcfs[[1]])

}
\seealso{
\code{\link{read_vcfs_as_granges}}
\code{\link{mut_type}}
}
