% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_remove_muttypes.R
\name{not_gr_or_grl}
\alias{not_gr_or_grl}
\title{Throw an error for when an argument should be a GRanges/GRangesList object.}
\usage{
not_gr_or_grl(arg)
}
\arguments{
\item{arg}{Argument. Any object that should have been a GRanges/GrangesList, but isn't.}
}
\description{
Throw an error for when an argument should be a GRanges/GRangesList object.
}
\details{
This function is called by other functions, when an argument should be a GRanges/GrangesList object, but isn't.
It throws an error showing the actual class of the argument.
}
\examples{
\dontrun{
a = 1
not_gr_or_grl(a)
}

}
