% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_96_spectrum.R
\name{plot_96_spectrum}
\alias{plot_96_spectrum}
\title{Plot 96 trinucleotide changes spectrum}
\usage{
plot_96_spectrum(mut_spectrum_96, colors = c("#DBD7C8", "#B2D39C", "#71C1BA",
  "#2DAFCE", "#2476B2", "#737E93"))
}
\arguments{
\item{mut_spectrum_96}{96 trinucleotide changes count matrix}
}
\value{
plot 96 trinucleotide changes spectrum plot
}
\description{
Plot 96 trinucleotide changes spectrum
}

