% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_context_heatmap.R
\name{plot_context_heatmap}
\alias{plot_context_heatmap}
\title{Plot mutation context heatmap}
\usage{
plot_context_heatmap(vcf_files, output)
}
\arguments{
\item{vcf_files}{Character vector containing vcf files}
}
\description{
Creates a heatmap of base substitution occurences with their context, normalized per sample and for trinucloetide occurences in the genome
}

