% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cosine_heatmap.R
\name{plot_cosine_heatmap}
\alias{plot_cosine_heatmap}
\title{Plot cosine similarity heatmap}
\usage{
plot_cosine_heatmap(
  cos_sim_matrix,
  col_order = NA,
  cluster_rows = TRUE,
  cluster_cols = FALSE,
  method = "complete",
  plot_values = FALSE
)
}
\arguments{
\item{cos_sim_matrix}{Matrix with pairwise cosine similarities.
Result from \code{\link{cos_sim_matrix}}}

\item{col_order}{Character vector with the desired order of the columns names for plotting. Optional.}

\item{cluster_rows}{Hierarchically cluster rows based on eucledian distance. Default = TRUE.}

\item{cluster_cols}{Hierarchically cluster cols based on eucledian distance. Default = FALSE.}

\item{method}{The agglomeration method to be used for hierarchical clustering. This should be one of 
"ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) 
or "centroid" (= UPGMC). Default = "complete".}

\item{plot_values}{Plot cosine similarity values in heatmap. Default = FALSE.}
}
\value{
Heatmap with cosine similarities
}
\description{
Plot pairwise cosine similarities in a heatmap.
}
\examples{

## See the 'mut_matrix()' example for how we obtained the mutation matrix:
mut_mat <- readRDS(system.file("states/mut_mat_data.rds",
                    package="MutationalPatterns"))

# You can download the signatures from the COSMIC website:
# https://cancer.sanger.ac.uk/cosmic/signature
## We copied the file into our package for your convenience.
filename <- system.file("extdata/snv_signatures_probabilities.txt",
                     package="MutationalPatterns")
signatures <- read.table(filename, sep = "\t", header = TRUE)
signatures = as.matrix(signatures[,-c(1,2)])

## Calculate the cosine similarity between each signature and each 96 mutational profile
cos_matrix = cos_sim_matrix(mut_mat, signatures)


## Plot the cosine similarity between each signature and each sample with hierarchical 
## sample clustering and signatures order based on similarity

plot_cosine_heatmap(cos_matrix, cluster_rows = TRUE, cluster_cols = TRUE, method = "complete")

## You can also plot the similarity of samples with eachother
cos_matrix = cos_sim_matrix(mut_mat, mut_mat)
plot_cosine_heatmap(cos_matrix, cluster_rows = TRUE, cluster_cols = TRUE, method = "complete")

}
\seealso{
\code{\link{mut_matrix}},
\code{\link{cos_sim_matrix}},
\code{\link{cluster_signatures}}
}
