% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lesion_segregation.R
\name{plot_lesion_segregation_gg}
\alias{plot_lesion_segregation_gg}
\title{Plot the strands of variants to show lesion segregation}
\usage{
plot_lesion_segregation_gg(
  tb,
  tb_limits,
  x_axis_breaks,
  point_size,
  sample_name
)
}
\arguments{
\item{tb}{A tibble with strand information of variants}

\item{tb_limits}{tibble describing the chromosome limits.
This ensures entire chromosomes are plotted,
instead of just the part with variants}

\item{x_axis_breaks}{x_axis_breaks}

\item{point_size}{Scalar describing the point size of the plot}

\item{sample_name}{Name of the sample}
}
\value{
ggplot2 object
}
\description{
This is a helper function for 'plot_lesion_segregation'.
It performs the actual plotting.
}
\seealso{
Other Lesion_segregation: 
\code{\link{calculate_lesion_segregation_gr}()},
\code{\link{calculate_lesion_segregation}()},
\code{\link{calculate_strand_switches}()},
\code{\link{calculate_strand_switch}()},
\code{\link{get_strandedness_gr}()},
\code{\link{get_strandedness_tb}()},
\code{\link{lesion_get_x_axis_breaks}()},
\code{\link{plot_lesion_segregation}()}
}
\concept{Lesion_segregation}
