% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profiles.R
\name{plot_profiles}
\alias{plot_profiles}
\title{Plot profiles for mutations found in mutation matrix}
\usage{
plot_profiles(mut_matrix, colors, ymax, type, method = "split",
  condensed = FALSE)
}
\arguments{
\item{mut_matrix}{Name list of mutation count matrices for mutation types.}

\item{colors}{(Optional) Named list with 6 value color vector "snv" for snv, 10 value color vector "dbs" for dbs.
For indels give same number of colors as there are classes. \cr
The "predefined" context contains 6 classes and the "cosmic" context contains 16 classes}

\item{ymax}{(Optional) Numeric vector for Y axis maximum value, order is snv, dbs, indel. Can be named
vector to specify mutation types. When "method = combine", "ymax" is the maximum of the
given values.\cr
Set "ymax = maximum" to use the maximal signature contribution for each mutation type as maximum
for the Y axis.\cr
By default, "ymax" is c("snv"=0.2, "dbs"=0.5, "indel"=0.5)}

\item{type}{(Optional) A character vector stating which type of mutation is to be extracted: 
'snv', 'dbs' and/or 'indel'. All mutation types can also be chosen by 'type = all'.\cr
Default is 'snv'}

\item{method}{(Optional) Character stating how to use the data. 
\itemize{
  \item{"split":} { Each mutation type has seperate profile plot}
  \item{"combine":} { Combined profile plots of all mutation types}
}   
Default is "split"}

\item{condensed}{(Optional) More condensed plotting format. \cr 
Default = FALSE}
}
\value{
The profile plot(s) of mutations for the given mutation types
}
\description{
Plot relative contribution of mutational profiles
}
\examples{
## See the 'mut_matrix_stranded()' example for how we obtained the
## mutation matrix with transcriptional strand information:
mut_mat <- readRDS(system.file("states/mut_mat_data.rds",
                                package="MutationalPatterns"))

## Plot the profiles of three samples
plot_profiles(mut_mat[,c(1,4,7)])

}
\seealso{
\code{\link{mut_matrix}}
}
