% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regional_similarity.R
\name{plot_regional_similarity}
\alias{plot_regional_similarity}
\title{Plot regional similarity}
\usage{
plot_regional_similarity(
  region_cossim,
  per_chrom = FALSE,
  oligo_correction = TRUE,
  max_cossim = NA
)
}
\arguments{
\item{region_cossim}{A region_cossim object.}

\item{per_chrom}{Boolean. Determines whether to create a separate plot per chromosome. (Default: FALSE)}

\item{oligo_correction}{Boolean describing whether the oligonucleotide
frequency corrected cosine similarities should be plotted. If no correction
has been applied then the regular cosine similarities will be plotted.
(Default: TRUE)}

\item{max_cossim}{Maximum cosine similarity for a window to be considered
an outlier. Any window with a lower cosine similarity is given a different
color. (Default: NA)}
}
\value{
ggplot2 object
}
\description{
Plot the cosine similarity of the mutation profiles of small genomic windows
with the rest of the genome.
}
\details{
Each dot shows the cosine similarity between the mutation profiles of a
single window and the rest of the genome. The dots are colored based on the
sizes in mega bases of the windows. This size is the distance between the
first and last mutations of a window. The locations of the mutations are
plotted on the bottom of the figure. The cosine similarity can be plotted
both with and without oligonucleotide frequency correction. This can be done
for all chromosomes at once or separate plots can be made per chromosome.
}
\examples{

## See the 'determine_regional_similarity()' example for how we obtained the
## following data:
regional_sims <- readRDS(system.file("states/regional_sims.rds",
  package = "MutationalPatterns"
))

## Plot the regional similarity
plot_regional_similarity(regional_sims)

## Plot outlier samples with a different color.
plot_regional_similarity(regional_sims, max_cossim = 0.5)

## Plot samples per chromosome
fig_l = plot_regional_similarity(regional_sims, per_chrom = TRUE)

}
\seealso{
\code{\link{determine_regional_similarity}}

Other regional_similarity: 
\code{\link{determine_regional_similarity}()}
}
\concept{regional_similarity}
