% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signatures_192.R
\name{plot_signatures_192}
\alias{plot_signatures_192}
\title{Plot 192 trinucleotide signatures}
\usage{
plot_signatures_192(signatures)
}
\arguments{
\item{signatures}{Dataframe containing mutational sigantures}
}
\description{
Plot 192 trinucleotide signatures
}

