% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signatures_288.R
\name{plot_signatures_288}
\alias{plot_signatures_288}
\title{Plot 288 trinucleotide signatures}
\usage{
plot_signatures_288(signatures)
}
\arguments{
\item{signatures}{Dataframe containing mutational sigantures}
}
\description{
Plot 288 trinucleotide signatures
}

