% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signatures_96.R
\name{plot_signatures_96}
\alias{plot_signatures_96}
\title{Plot 96 trinucleotide signatures}
\usage{
plot_signatures_96(signatures)
}
\arguments{
\item{vcf_files}{Character vector of vcf_files}

\item{rank}{Number of signatures one would like to extract}

\item{outdir}{Specify output directory}
}
\description{
Plot 96 trinucleotide signatures
}

