% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_to_signatures_strict.R
\name{plot_sim_decay}
\alias{plot_sim_decay}
\title{Plot decay in cosine similarity as signatures are removed.}
\usage{
plot_sim_decay(sims, removed_sigs, max_delta)
}
\arguments{
\item{sims}{List of cosine similarities}

\item{removed_sigs}{List of iteratively removed signatures}

\item{max_delta}{The maximum difference in original vs reconstructed cosine similarity.}
}
\value{
ggplot object
}
\description{
This function is called by fit_to_signatures_strict
}
