% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum_region.R
\name{plot_spectrum_region}
\alias{plot_spectrum_region}
\title{Plot point mutation spectrum per genomic region}
\usage{
plot_spectrum_region(
  type_occurrences,
  by = NA,
  mode = c("relative", "absolute"),
  colors = NULL,
  legend = TRUE
)
}
\arguments{
\item{type_occurrences}{Type occurrences matrix}

\item{by}{Optional grouping variable}

\item{mode}{'absolute' or 'relative'.
When relative, the number of variants will be shown
divided by the total number of variants in that sample and genomic region.}

\item{colors}{Optional color vector with 7 values}

\item{legend}{Plot legend, default = TRUE}
}
\value{
Spectrum plot by genomic region
}
\description{
A spectrum similar to the one from 'plot_spectrum()' is plotted.
However the spectrum is plotted separately per genomic region.
As input it takes a 'type_occurrences' matrix that was calculated per genomic region.
To get a 'type_occurrences' matrix per region,
first use the 'split_muts_region()' function on a GR or GRangesList.
Then use the 'mut_type_occurrences' as you would normally.
The by, colors and legend argument work the same as in'plot_spectrum()'.
}
\examples{
## See the 'split_muts_region()' example for how we obtained the
## following data:
grl <- readRDS(system.file("states/grl_split_region.rds",
                package="MutationalPatterns"))

## Or plot spectrum per tissue
tissue <- c("colon", "colon", "colon",
            "intestine", "intestine", "intestine",
            "liver", "liver", "liver")

 ## Load a reference genome.
ref_genome = "BSgenome.Hsapiens.UCSC.hg19"
library(ref_genome, character.only = TRUE)


## Get the type occurrences for all VCF objects.
type_occurrences = mut_type_occurrences(grl, ref_genome)

## Plot the point mutation spectrum per sample type and per genomic region
plot_spectrum_region(type_occurrences, by = tissue)

## Plot the absolute point mutation spectrum per sample type and per genomic region
plot_spectrum_region(type_occurrences, mode = "absolute")

sample_names <- c(
"colon1", "colon2", "colon3",
"intestine1", "intestine2", "intestine3",
"liver1", "liver2", "liver3")

## Plot the point mutation spectrum per individual sample and per genomic region
plot_spectrum_region(type_occurrences, by = sample_names)


}
\seealso{
\code{\link{read_vcfs_as_granges}},
\code{\link{mut_type_occurrences}},
\code{\link{plot_spectrum}}
}
