% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_strand_profiles.R
\name{plot_strand_profiles}
\alias{plot_strand_profiles}
\title{Plot transcription/replication profiles}
\usage{
plot_strand_profiles(mut_matrix, colors, ymax, type, mode,
  method = "split", condensed = FALSE)
}
\arguments{
\item{mut_matrix}{Named list of mutation count matrices for mutation types 
with transcriptional or replication strand information}

\item{colors}{(Optional) Named list with 6 value color vector "snv" for snv, 10 value color vector "dbs" for dbs.
For indels give same number of colors as there are classes. \cr
The "predefined" context contains 6 classes and the "cosmic" context contains 16 classes}

\item{ymax}{(Optional) Numeric vector for Y axis maximum value, order is snv, dbs, indel. Can be named
vector to specify mutation types. When "method = combine", "ymax" is the maximum of the
given values.\cr
Set "ymax = maximum" to use the maximal signature contribution for each mutation type as maximum
for the Y axis.\cr
By default, "ymax" is c("snv"=0.2, "dbs"=0.5, "indel"=0.5)}

\item{type}{(Optional) A character vector stating which type of mutation is to be extracted: 
'snv', 'dbs' and/or 'indel'. All mutation types can also be chosen by 'type = all'.\cr
Default is 'snv'}

\item{mode}{Character stating which type of process to look at: transcription or replication}

\item{method}{(Optional) Character stating how to use the data. 
\itemize{
  \item{"split":} { Each mutation type has seperate profile plot}
  \item{"combine":} { Combined profile plots of all mutation types}
}   
Default is "split"}

\item{condensed}{(Optional) More condensed plotting format. \cr 
Default = FALSE}
}
\value{
The profile plot(s) of mutations for the given mutation types with transcriptional or
replication strand information
}
\description{
Plot mutational profiles with transcription or replication strand information
}
\examples{
## See the 'mut_matrix_stranded()' example for how we obtained the
## mutation matrix with transcriptional strand information:
mut_mat_s <- readRDS(system.file("states/mut_mat_s_data.rds",
                                    package="MutationalPatterns"))

## Extract the signatures.
## This is a computationally intensive task, so we load a precomputed
## version instead.
# nmf_res_strand <- extract_signatures(mut_mat_s, rank = 2)
nmf_res_strand <- readRDS(system.file("states/nmf_res_strand_data.rds",
                                    package="MutationalPatterns"))

## Optionally, provide signature names
colnames(nmf_res_strand$signatures) <- c("Signature A", "Signature B")

## Generate the plot
plot_strand_profiles(nmf_res_strand$signatures)

}
\seealso{
\code{\link{mut_matrix_stranded}},
\code{\link{extract_signatures}}
}
