% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_bootstrap_comparison.R
\name{profile_bootstrap_comparison}
\alias{profile_bootstrap_comparison}
\title{compare two profiles by bootstrapping}
\usage{
profile_bootstrap_comparison(profile1, profile2, random.seed = NULL,
  n = 500, reportSimultations = FALSE)
}
\arguments{
\item{profile1}{A numeric vector giving the frequencies of each mutation type.}

\item{profile2}{A second profile to compare the first one two.}

\item{random.seed}{Optionally pass on a random seed to make results reproducable.}

\item{n}{The number of iterations.}

\item{reportSimulations}{Boolean indicating wether to report all results or not.}
}
\value{
A list with the elements \code{overallPvalue} giving a p-value for the overal comparison,
        \code{mutTypePvalues}, giving the p-values by mutation type and \code{mutTypePvalues_corrected},
        giving the p-values by mutation type corrected for multiple testing. If the option
        \code{reportSimulations} is set, there will be an additional element \code{simulatedProfileDifferences},
        giving a matrix of n times the number of mutation types values indicating the profile differences per simulation.
}
\description{
Get a p-value for the differences between two profiles.
}

