% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcfs_as_granges.R
\name{read_single_vcf_as_grange}
\alias{read_single_vcf_as_grange}
\title{Read a single VCF file into a GRanges object}
\usage{
read_single_vcf_as_grange(vcf_file, genome, group, change_seqnames)
}
\arguments{
\item{vcf_file}{A VCF file name}

\item{genome}{BSgenome object}

\item{group}{Selector for a seqlevel group.  All seqlevels outside
of this group will be removed.  Possible values:
* 'all' for all chromosomes;
* 'auto' for autosomal chromosomes;
* 'sex' for sex chromosomes;
* 'auto+sex' for autosomal + sex chromosomes (default);
* 'circular' for circular chromosomes;
* 'none' for no filtering, which results in keeping all
  seqlevels from the VCF file.}

\item{change_seqnames}{Boolean. Whether to change the seqnamesStyle of the vcf
to that of the BSgenome object.}
}
\value{
A GRanges object
}
\description{
This function reads a Variant Call Format (VCF) file into a GRanges object
In addition to loading the files, this
function applies the same seqlevel style to the GRanges objects as the
reference genome passed in the 'genome' parameter.
}
