% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf.R
\name{read_vcf}
\alias{read_vcf}
\title{Read vcf files into list of CollapsedVCF objects}
\usage{
read_vcf(vcf_files, sample_names, genome = "-")
}
\arguments{
\item{vcf_files}{Character vector of vcf file names}

\item{sample_names}{Character vector of sample names}
}
\value{
List of CollapsedVCF objects
}
\description{
Function reads Variant Call Format (VCF) files into a CollapsedVCF object and combines them in a list object
}
\examples{
vcf_file_list = list.files("/your_vcf_dir/", full.names = T) 
vcf_list = read_vcf_list(vcf_files_list, sample_names, genome = "hg19")
}

