% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf_list.R
\name{read_vcf_list}
\alias{read_vcf_list}
\title{Read vcf files into list of CollapsedVCF objects}
\usage{
read_vcf_list(vcf_file_list, sample_names)
}
\arguments{
\item{vcf_file_list}{Character vector of vcf file names}

\item{sample_names}{Character vector of sample names}
}
\value{
List of CollapsedVCF objects
}
\description{
Function reads Variant Call Format (VCF) files into a CollapsedVCF object and combines them in a list object
}
\examples{
vcf_file_list = list.files("/your_vcf_dir/", full.names = T) 
vcf_list = read_vcf_list(vcf_files_list, sample_names)
}

