% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_remove_muttypes.R
\name{remove_indels}
\alias{remove_indels}
\title{Removes Indel variants.}
\usage{
remove_indels(grl)
}
\arguments{
\item{grl}{GRanges/GRangesList object}
}
\value{
A filtered version of the input GRanges/GrangesList object.
}
\description{
Removes Indel variants.
}
\details{
This function removes Indel variants for a GRanges/GrangesList object.
}
