% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_remove_muttypes.R
\name{remove_indels_gr}
\alias{remove_indels_gr}
\title{Removes Indel variants.}
\usage{
remove_indels_gr(gr)
}
\arguments{
\item{gr}{GRanges object}
}
\value{
A filtered version of the input GRanges object.
}
\description{
Removes Indel variants.
}
\details{
This function removes Indel variants for a single GRanges object.
}
