% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_remove_muttypes.R
\name{remove_multi_alts_variants}
\alias{remove_multi_alts_variants}
\title{Removes variants with multiple alt alleles.}
\usage{
remove_multi_alts_variants(grl)
}
\arguments{
\item{grl}{GRanges/GRangesList object}
}
\value{
A filtered version of the input GRanges/GRangesList object.
}
\description{
Removes variants with multiple alt alleles.
}
\details{
This function removes variants with multiple alternative alleles for a GRanges/GRangesList object.
}
\examples{

}
