% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_remove_muttypes.R
\name{remove_multi_alts_variants_gr}
\alias{remove_multi_alts_variants_gr}
\title{Removes variants with multiple alt alleles.}
\usage{
remove_multi_alts_variants_gr(gr)
}
\arguments{
\item{gr}{GRanges object}
}
\value{
A filtered version of the input GRanges object.
}
\description{
Removes variants with multiple alt alleles.
}
\details{
This function removes variants with multiple alternative alleles for a single GRanges object.
}
