% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_remove_muttypes.R
\name{remove_snvs}
\alias{remove_snvs}
\title{Removes SNV/MNV variants.}
\usage{
remove_snvs(grl)
}
\arguments{
\item{grl}{GRanges/GRangesList object}
}
\value{
A filtered version of the input GRanges/GrangesList object.
}
\description{
Removes SNV/MNV variants.
}
\details{
This function removes SNV/MNV variants for a GRanges/GrangesList object.
}
\examples{

}
