% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_remove_muttypes.R
\name{remove_snvs_gr}
\alias{remove_snvs_gr}
\title{Removes SNV/MNV variants.}
\usage{
remove_snvs_gr(gr)
}
\arguments{
\item{gr}{GRanges object}
}
\value{
A filtered version of the input GRanges object.
}
\description{
Removes SNV/MNV variants.
}
\details{
This function removes SNV/MNV variants for a single GRanges object.
}
\examples{

}
