% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_chrom.R
\name{rename_chrom}
\alias{rename_chrom}
\title{Rename chromosome names to standard}
\usage{
rename_chrom(granges, style = "UCSC")
}
\arguments{
\item{granges}{Granges object with chromosome names to rename}

\item{style}{"UCSC", NCBI" or "Ensembl", default = "UCSC"}
}
\value{
A GRanges object with renamed chromosomes according to style.
}
\description{
Rename chromosome names to standard
}
\examples{
## After loading VCF files using 'vcf_to_granges()', the chromosome names
## may differ between the reference genome and your samples.  See the
## 'vcf_to_granges()' example for how we obtain the following data.
vcfs <- readRDS(system.file("states/vcf_to_granges_output.R",
                package="MutationalPatterns"))

## You can standardize the naming of your VCF samples loaded into R.
## Notice how the seqnames change from:
vcfs$liver1

## To the UCSC (default) standardized format:
rename_chrom(vcfs$liver1)

## You can also rename all objects returned by 'vcf_to_granges()' using:
vcfs <- lapply(vcfs, rename_chrom)

}
\seealso{
\code{\link{vcf_to_granges}}
}

