% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_to_signatures_bootstrapped.R
\name{resample_mut_mat}
\alias{resample_mut_mat}
\title{Resample a mutation matrix}
\usage{
resample_mut_mat(mut_matrix)
}
\arguments{
\item{mut_matrix}{mutation count matrix (dimensions: x mutation types
X n samples)}
}
\value{
A resamples mutation matrix
}
\description{
The mutation matrix is resampled per column (sample).
Resampling is done with replacement using the row weights as propabilities.
}
\seealso{
\code{\link{fit_to_signatures_bootstrapped}}
}
