% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_context_dbs.R
\name{set_context_dbs}
\alias{set_context_dbs}
\title{Set DBS context}
\usage{
set_context_dbs(grl)
}
\arguments{
\item{grl}{GRanges/GRangesList}
}
\value{
A version of the GRanges/GRangesList object, with modified REF and ALT columns.
}
\description{
Set the DBS COSMIC context on an GRanges/GRangesList object.
It applies the get_indel_context_gr function to each gr in the input,
which works by changing the REF and ALT columns of the GRanges into the COSMIC types.
}
\examples{
## Get GRangesList with DBS.
## See 'get_mut_type' or 'read_vcfs_as_granges' for more info on how to do this.
grl_dbs <- readRDS(system.file("states/blood_grl_dbs.rds",
                package="MutationalPatterns"))
                
##Set context dbs
set_context_dbs(grl_dbs)

}
\seealso{
\code{\link{set_context_dbs_gr}}, \code{\link{get_mut_type}}, \code{\link{read_vcfs_as_granges}}
}
