% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbs_set_context.R
\name{set_dbs_context}
\alias{set_dbs_context}
\title{Set DBS context}
\usage{
set_dbs_context(grl)
}
\arguments{
\item{grl}{GRanges/GRangesList}
}
\value{
A version of the GRanges/GRangesList object, with modified REF and ALT columns.
}
\description{
Set the DBS COSMIC context on an GRanges/GRangesList object.
It applies the get_indel_context_gr function to each gr in the input,
which works by changing the REF and ALT columns of the GRanges into the COSMIC types.
}
\examples{
## Get GRangesList with DBS.
## See 'get_mut_type' or 'read_vcfs_as_granges' for more info on how to do this.
grl_dbs <- readRDS(system.file("states/blood_grl_dbs.rds",
                package="MutationalPatterns"))
                
##Set context dbs
set_dbs_context(grl_dbs)

}
\seealso{
\code{\link{set_dbs_context_gr}}, \code{\link{get_mut_type}}, \code{\link{read_vcfs_as_granges}}

Other DBS: 
\code{\link{count_dbs_contexts_gr}()},
\code{\link{count_dbs_contexts}()},
\code{\link{plot_dbs_contexts}()},
\code{\link{plot_main_dbs_contexts}()},
\code{\link{set_dbs_context_gr}()}
}
\concept{DBS}
