% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_class.R
\name{show,region_cossim-method}
\alias{show,region_cossim-method}
\title{An S4 method to show an instance of the region_cossim class.}
\usage{
\S4method{show}{region_cossim}(object)
}
\arguments{
\item{object}{A region_cossim object.}
}
\description{
An S4 method to show an instance of the region_cossim class.
}
