% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mut_type.R
\name{split_mbs_gr}
\alias{split_mbs_gr}
\title{Split SNV/MNV into SNV, DBS and MNVs of different sizes}
\usage{
split_mbs_gr(gr, merge_muts = T)
}
\arguments{
\item{gr}{A GRanges object}

\item{merge_muts}{Boolean value. If TRUE, the mutations are merged.}
}
\value{
A list of granges
}
\description{
A function that splits a GRanges object into a list of GRanges.
Mutations are split based on whether they are a SNV/MNV. They are split based on their size.
}
\details{
This function assumes that a DBS/MNV is called as separate SNVs, that are located next to each other.
This function works by checking the location of these SNVs. 
If multiple SNVs are located next to eachother they are considered a DBS/MNV.
A CAT>TAG would be considered as a MNV of length 3.
The largest variants are then selected and put into a separate GRanges object.
This is done recursively untill all variants are separated based on their size.
This way the CAT>TAG mutation will not be seen as two separate DBS variants.
This results in a list of granges.
If merge_muts is TRUE, then the SNVs that make up a DBS/MNV will be merged.
}
\seealso{
\code{\link{merge_muts}}, \code{\link{get_mut_type_gr}}
}
