% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_muts_region.R
\name{split_muts_region}
\alias{split_muts_region}
\title{Split GRangesList or GRanges based on a list of regions.}
\usage{
split_muts_region(grl, ranges_grl)
}
\arguments{
\item{grl}{GRangesList or GRanges object}

\item{ranges_grl}{GRangesList or GRanges object containing regions of interest}
}
\value{
GRangesList
}
\description{
A GRangesList or GRanges object containing variants is split based on a list of regions.
This list can be either a GRangesList or a GRanges object.
The result is a GRangesList where each element contains the variants of one sample from one region.
Variant that are not in any of the provided region are put in a list of 'other'.
}
\examples{

## Read in some existing genomic regions. 
## See the 'genomic_distribution()' example for how we obtained the
## following data:
CTCF_g <- readRDS(system.file("states/CTCF_g_data.rds",
                    package="MutationalPatterns"))
promoter_g <- readRDS(system.file("states/promoter_g_data.rds",
                        package="MutationalPatterns"))
flanking_g <- readRDS(system.file("states/promoter_flanking_g_data.rds",
                                    package="MutationalPatterns"))

## Combine the regions into a single GRangesList
regions <- GRangesList(promoter_g, flanking_g, CTCF_g)

names(regions) <- c("Promoter", "Promoter flanking", "CTCF")

## Read in some variants.
## See the 'read_vcfs_as_granges()' example for how we obtained the
## following data:
grl <- readRDS(system.file("states/read_vcfs_as_granges_output.rds",
                package="MutationalPatterns"))
                
split_muts_region(grl, regions)


}
\seealso{
Other genomic_regions: 
\code{\link{get_muts_region}()},
\code{\link{split_muts_region_gr}()}
}
\concept{genomic_regions}
