% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_muts_region.R
\name{split_muts_region_gr}
\alias{split_muts_region_gr}
\title{Split a GRanges object based on a list of regions.}
\usage{
split_muts_region_gr(gr, ranges_grl)
}
\arguments{
\item{gr}{GRanges object containing variants}

\item{ranges_grl}{GRangesList or GRanges object containing regions of interest}
}
\value{
GRangesList
}
\description{
A GRanges object containing variants is split based on a list of regions.
This list can be either a GRangesList or a GRanges object.
The result is a GRangesList where each element contains the variants of one sample from one region.
Variant that are not in any of the provided region are put in a list of 'other'.
}
\seealso{
Other genomic_regions: 
\code{\link{get_muts_region}()},
\code{\link{split_muts_region}()}
}
\concept{genomic_regions}
