% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_strand.R
\name{strand_from_vcf}
\alias{strand_from_vcf}
\title{This function has been renamed. Use 'mut_strand' instead.}
\usage{
strand_from_vcf(vcf, genes)
}
\arguments{
\item{vcf}{A GRanges object}

\item{genes}{The genes}
}
\value{
Character vector with transcriptional strand information
}
\description{
This function has been renamed. Use 'mut_strand' instead.
}
\seealso{
\code{\link{mut_strand}}
}
