% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_to_granges.R
\name{bed_to_granges}
\alias{bed_to_granges}
\title{Read bed files as a list of GRanges objects}
\usage{
bed_to_granges(bed_files, region_names)
}
\arguments{
\item{bed_files}{Character vector with bed files}

\item{region_names}{Character vector with names of regions in bed files}
}
\value{
granges_list List of Granges objects
}
\description{
Read bed files as a list of GRanges objects
}
\examples{
surveyed_file = list.files(system.file("extdata",
                                       package="MutationalPatterns"),
                           pattern = ".bed",
                           full.names = TRUE)
surveyed_list = bed_to_granges(surveyed_file, "surveyed_all")

}
\seealso{
\code{\link{genomic_distribution}}
}

