% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_to_granges.R
\name{bed_to_granges}
\alias{bed_to_granges}
\title{Read region bed files}
\usage{
bed_to_granges(bed_files, region_names)
}
\arguments{
\item{bed_files}{Character vector with bed files}

\item{region_names}{Character vector with names of regions}
}
\value{
granges_list List of Granges objects
}
\description{
Read region bed files as a list of Granges objects
}

