% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial_test.R
\name{binomial_test}
\alias{binomial_test}
\title{Binomial test for enrichment or depletion testing}
\usage{
binomial_test(prob, surveyed_region_length, observed)
}
\arguments{
\item{surveyed_region_length}{Number of bp that is surveyed in region}

\item{observed}{Observed number of mutations in surveyed region}

\item{n_muts}{Number of mutations in whole surveyed genome}

\item{surveyed_length}{Number of bp that was surveyed in whole genome}
}
\description{
Test whether specific genomic region is depleted or enriched using a binomial test
}

