% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lesion_segretation.R
\name{calculate_lesion_segregation}
\alias{calculate_lesion_segregation}
\title{Calculate the amount of lesion segregation for a GRangesList or GRanges object.}
\usage{
calculate_lesion_segregation(
  grl,
  sample_names,
  split_by_type = FALSE,
  ref_genome = NA
)
}
\arguments{
\item{grl}{GRangesList or GRanges object}

\item{sample_names}{The name of the sample}

\item{split_by_type}{Boolean describing whether the lesion 
segregation should be calculated for all SNVs together or per 96 substitution context}

\item{ref_genome}{A string matching the name of a BSgenome library
corresponding to the reference genome.
Only needed when split_by_type is TRUE}
}
\value{
A tibble containing the amount of lesions segregation per sample
}
\description{
The amount of lesion segregation is calculated per GRanges object.
The results are then combined in a table.
It's possible to calculate the lesion segregation separately per 96 substitution context.
The results are then automatically added back up together.
}
\examples{

## See the 'read_vcfs_as_granges()' example for how we obtained the
## following data:
grl <- readRDS(system.file("states/read_vcfs_as_granges_output.rds",
                package="MutationalPatterns"))

## Set the sample names
sample_names <- c(
   "colon1", "colon2", "colon3",
   "intestine1", "intestine2", "intestine3",
   "liver1", "liver2", "liver3")
                                 
## Load the corresponding reference genome.
ref_genome = "BSgenome.Hsapiens.UCSC.hg19"
library(ref_genome, character.only = TRUE)

## Calculate lesion segregation
lesion_segretation = calculate_lesion_segregation(grl, sample_names)

## Calculate lesion segregation per 96 base type
lesion_segretation_by_type = calculate_lesion_segregation(grl, sample_names, 
split_by_type = TRUE, ref_genome = ref_genome)

}
\seealso{
\code{\link{plot_lesion_segregation}}

Other Lesion_segregation: 
\code{\link{calculate_lesion_segregation_gr}()},
\code{\link{calculate_strand_switches}()},
\code{\link{calculate_strand_switch}()},
\code{\link{get_strandedness_gr}()},
\code{\link{get_strandedness_tb}()},
\code{\link{lesion_get_x_axis_breaks}()},
\code{\link{plot_lesion_segregation_gg}()},
\code{\link{plot_lesion_segregation}()}
}
\concept{Lesion_segregation}
