% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lesion_segretation.R
\name{calculate_strand_switch}
\alias{calculate_strand_switch}
\title{Calculate the number of strand switches in a single chromosome.}
\usage{
calculate_strand_switch(strand)
}
\arguments{
\item{strand}{A vector containing the strand of variants}
}
\value{
A boolean vector describing for each variant 
if it switched strands with the previous variant.
}
\description{
Calculate the number of strand switches in a single chromosome.
}
\seealso{
Other Lesion_segregation: 
\code{\link{calculate_lesion_segregation_gr}()},
\code{\link{calculate_lesion_segregation}()},
\code{\link{calculate_strand_switches}()},
\code{\link{get_strandedness_gr}()},
\code{\link{get_strandedness_tb}()},
\code{\link{lesion_get_x_axis_breaks}()},
\code{\link{plot_lesion_segregation_gg}()},
\code{\link{plot_lesion_segregation}()}
}
\concept{Lesion_segregation}
