% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lesion_segretation.R
\name{calculate_strand_switches}
\alias{calculate_strand_switches}
\title{Calculate the total number of variants and strand switches.}
\usage{
calculate_strand_switches(tb)
}
\arguments{
\item{tb}{A tibble with strand information}
}
\value{
A list containing the total number of variants and the number of strand switches
}
\description{
Calculate the total number of variants and strand switches.
}
\seealso{
Other Lesion_segregation: 
\code{\link{calculate_lesion_segregation_gr}()},
\code{\link{calculate_lesion_segregation}()},
\code{\link{calculate_strand_switch}()},
\code{\link{get_strandedness_gr}()},
\code{\link{get_strandedness_tb}()},
\code{\link{lesion_get_x_axis_breaks}()},
\code{\link{plot_lesion_segregation_gg}()},
\code{\link{plot_lesion_segregation}()}
}
\concept{Lesion_segregation}
