% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos_sim.R
\name{cos_sim}
\alias{cos_sim}
\title{Cosine similarity function}
\usage{
cos_sim(x, y)
}
\arguments{
\item{x}{Vector 1}

\item{y}{Vector 2}
}
\value{
Cosine similarity value
}
\description{
Calculate the cosine similarity between two vectors
}
\examples{

}
