% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cos_sim_matrix.R
\name{cos_sim_matrix}
\alias{cos_sim_matrix}
\title{Creat a mutational matrix with cosine similarities}
\usage{
cos_sim_matrix(mut_matrix, cancer_signatures)
}
\arguments{
\item{mut_matrix1}{96 mutation count matrix (dimensions: 96 mutations X n samples)}

\item{mut_matrix2}{96 mutation count matrix (dimensions: 96 mutations X m samples)}
}
\value{
Matrix with pairwise cosine similarities (dimensions: n mutational profiles X m mutational profiles)
}
\description{
Calculates the pairwise cosine similarity between the mutational profiles provided in the two mutation count matrices. 
The cosine similarity is a value between 0 (distinct) and 1 (identical) and indicates how much two vectors are alike.
}
\examples{
## You can download the signatures from the pan-cancer study by
## Alexandrov et al:
# http://cancer.sanger.ac.uk/cancergenome/assets/signatures_probabilities.txt

## We copied the file into our package for your convenience.

filename <- system.file("extdata/signatures_probabilities.txt",
                        package="MutationalPatterns")
cancer_signatures <- read.table(filename, sep = "\\t", header = TRUE)
## Reorder the columns to make the order of the trinucleotide changes compatible.
cancer_signatures <- cancer_signatures[order(cancer_signatures[,1]),]
## Include only the signatures in the matrix.
cancer_signatures <- as.matrix(cancer_signatures[,4:33])

## See the 'mut_matrix()' example for how we obtained the mutation matrix:
mut_mat <- readRDS(system.file("states/mut_mat_data.rds",
                    package="MutationalPatterns"))

## Calculate the cosine similarity between each signature and each 96 mutational profile

cos_sim_matrix(mut_mat, cancer_signatures)

}
\seealso{
\code{\link{mut_matrix}},
\code{\link{fit_to_signatures}},
\code{\link{plot_cosine_heatmap}}
}
