% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_depletion_test.R
\name{enrichment_depletion_test}
\alias{enrichment_depletion_test}
\title{Enrichment depletion test}
\usage{
enrichment_depletion_test(prob, observed, surveyed_region_length)
}
\arguments{
\item{observed}{Observed number of mutations in surveyed region}

\item{surveyed_region_length}{Number of bp that is surveyed in region}

\item{n_muts}{Number of mutations in whole surveyed genome}

\item{surveyed_length}{Number of bp that was surveyed in whole genome}
}
\description{
Test whether specific genomic region is depleted or enriched using binomial test
}

