% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_depletion_test.R
\name{enrichment_depletion_test}
\alias{enrichment_depletion_test}
\title{Find overlaps between mutations and a genomic region}
\usage{
enrichment_depletion_test(x, by = c())
}
\arguments{
\item{x}{Dataframe result from genomic_distribution()}

\item{by}{Optional aggregation vector}
}
\description{
aggregates (optional) and then performs enrichment depletion test
}

