% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_rank.R
\name{estimate_rank}
\alias{estimate_rank}
\title{Estimate optimal rank for NMF decomposition}
\usage{
estimate_rank(mut_matrix, rank_range, nrun = 100)
}
\arguments{
\item{mut_matrix}{96 mutation count matrix}

\item{rank_range}{Range of ranks one would like to test}

\item{nrun}{Number of runs to perform, default=100}
}
\value{
NMF rank survey plot
}
\description{
Find optimal number of signatures for NMF decomposition
}
\examples{
vcf_files = list.files(system.file("extdata", package="MutationalPatterns"),
                                   pattern = ".vcf",
                                   full.names = TRUE)
sample_names = c("colon1", "colon2", "colon3",
                 "intestine1", "intestine2", "intestine3",
                 "liver1", "liver2", "liver3")
vcfs = read_vcf(vcf_files, sample_names, genome = "hg19")

# only select autosomal chromosomes, mt dna length is different for vcf and
# ref genome, why??
auto = extractSeqlevelsByGroup(species="Homo_sapiens",
                               style="UCSC",
                               group="auto")
vcfs = lapply(vcfs, function(x) keepSeqlevels(x, auto))

ref_genome = "BSgenome.Hsapiens.UCSC.hg19" 
library(ref_genome, character.only = TRUE)
tissue = c("colon", "colon", "colon",
           "intestine", "intestine", "intestine",
           "liver", "liver", "liver")
vcfs = lapply(vcfs, function(x) rename_chrom(x))

test_matrix = mut_matrix(vcf_list = vcfs, ref_genome = ref_genome)
estimate_rank(test_matrix, rank_range = 2:5, nrun = 50)

}

