% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_rank.R
\name{estimate_rank}
\alias{estimate_rank}
\title{Estimate optimal rank for NMF decomposition}
\usage{
estimate_rank(mut_matrix, rank_range, nrun = 100)
}
\arguments{
\item{mut_matrix}{96 mutation count matrix}

\item{rank_range}{Range of ranks one would like to test}

\item{nrun}{Number of runs to perform, default=100}
}
\value{
NMF rank survey plot
}
\description{
Find optimal number of signatures for NMF decomposition
}
\examples{
## See the 'read_vcf()' example for how we obtained the following data:
vcfs <- readRDS(system.file("states/read_vcf_output.R",
                package="MutationalPatterns"))

# Rename the seqlevels to the UCSC standard.
vcfs <- lapply(vcfs, rename_chrom)

# Exclude mitochondrial and allosomal chromosomes.
autosomal = extractSeqlevelsByGroup(species="Homo_sapiens",
                                    style="UCSC",
                                    group="auto")

vcfs = lapply(vcfs, function(x) keepSeqlevels(x, autosomal))

## Define the reference genome we are going to use.
ref_genome = "BSgenome.Hsapiens.UCSC.hg19" 

## If neccessary, download it from Bioconductor.
# source("http://bioconductor.org/biocLite.R")
# biocLite(ref_genome)

## Load the reference genome.
library(ref_genome, character.only = TRUE)

test_matrix = mut_matrix(vcf_list = vcfs, ref_genome = ref_genome)
estimate_rank(test_matrix, rank_range = 2:5, nrun = 50)

}
\seealso{
\code{\link{read_vcf}}, \code{\link{rename_chrom}}
}

