% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explained_by_signatures.R
\name{explained_by_signatures}
\alias{explained_by_signatures}
\title{Determine how much of a mutational profile can be explained by a mutational signature.}
\usage{
explained_by_signatures(mut_matrix, cancer_signatures)
}
\arguments{
\item{mut_matrix}{96 mutation count matrix (dimensions: 96 mutations X n samples)}

\item{signatures}{Signature matrix (dimensions: 96 mutations X n signatures)}
}
\value{
Matrix with pairwise cosine similarities (dimensions: n samples X n signatures)
}
\description{
Calculates the cosine similarity for each pair of mutation profile and mutational signature. 
The cosine similarity is a value between 0 (distinct) and 1 (identical). 
The cosine similarity indicates how much of the 96 
mutation profile can be explained by an individual signature.
}
\examples{
## You can download the signatures from the pan-cancer study by
## Alexandrov et al:
# http://cancer.sanger.ac.uk/cancergenome/assets/signatures_probabilities.txt

## We copied the file into our package for your convenience.
filename <- system.file("extdata/signatures_probabilities.txt",
                        package="MutationalPatterns")

cancer_signatures <- read.table(filename, sep = "\\t", header = TRUE)

## Reorder the columns to make the order of the trinucleotide changes compatible.
cancer_signatures <- cancer_signatures[order(cancer_signatures[,1]),]

## Include only the signatures in the matrix.
cancer_signatures <- as.matrix(cancer_signatures[,4:33])

## See the 'mut_matrix()' example for how we obtained the mutation matrix:
mut_mat <- readRDS(system.file("states/mut_mat_data.rds",
                    package="MutationalPatterns"))

## Calculate the cosine similarity between each signature and each 96 mutational profile

explained_by_signatures(mut_mat, cancer_signatures)

}
\seealso{
\code{\link{mut_matrix}},
\code{\link{fit_to_signatures}}
}
