% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_signatures.R
\name{extract_signatures}
\alias{extract_signatures}
\title{Extract mutational signatures using NMF}
\usage{
extract_signatures(mut_spectrum, rank, nrun = 200)
}
\arguments{
\item{rank}{Number of signatures one would like to extract}

\item{nrun}{Number of iterations, default = 200}

\item{vcf_files}{Character vector of vcf_files}

\item{outdir}{Specify output directory}
}
\description{
Decomposes trinucleotide count matrix into signatures and contribution of those signatures to the spectra of the individuals samples/vcf files
}

