% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_nested.R
\docType{data}
\name{facet_nested}
\alias{facet_nested}
\alias{FacetNested}
\title{Plot merged facets}
\usage{
facet_nested(rows = NULL, cols = NULL, scales = "fixed",
  space = "fixed", shrink = TRUE, labeller = "label_value",
  as.table = TRUE, switch = NULL, drop = TRUE, margins = FALSE,
  facets = NULL, nest_line = FALSE, resect = unit(0, "mm"),
  bleed = FALSE)
}
\description{
Helper function for ggplot to merge facets
}
\keyword{datasets}
