% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_to_signatures.R
\name{fit_to_signatures}
\alias{fit_to_signatures}
\title{Find optimal nonnegative linear combination of mutation signatures to reconstruct the mutation matrix}
\usage{
fit_to_signatures(mut_matrix, signatures)
}
\arguments{
\item{mut_matrix}{96 mutation count matrix (dimensions: 96 mutations X n samples)}

\item{signatures}{Signature matrix (dimensions: 96 mutations X n signatures)}
}
\value{
Named list with signature contributions and reconstructed mutation matrix
}
\description{
Find linear combination of mutation signatures that most closely reconstructs the mutation matrix by solving nonnegative least-squares constraints problem
}

