% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_distribution.R
\name{genomic_distribution}
\alias{genomic_distribution}
\title{Find overlaps between mutations and a genomic region}
\usage{
genomic_distribution(vcf, surveyed, region)
}
\arguments{
\item{vcf}{CollapsedVCF object containing mutations}

\item{surveyed}{GRanges object containing regions of the genome that were surveyed}

\item{region}{GRanges object containing locations of genomic region(s)}
}
\description{
Function finds the number of mutations that reside in genomic region and takes surveyed area of genome into account
}

