% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_distribution.R
\name{genomic_distribution}
\alias{genomic_distribution}
\title{Find overlaps between mutations and a genomic region}
\usage{
genomic_distribution(vcf_list, surveyed_list, region_list)
}
\arguments{
\item{vcf_list}{A list...}

\item{region_list}{List with GRanges objects containing locations of genomic regions}
}
\description{
Function finds the number of mutations that reside in genomic region and takes surveyed area of genome into account
}

