% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_distribution.R
\name{genomic_distribution}
\alias{genomic_distribution}
\title{Find overlaps between mutations and a genomic region}
\usage{
genomic_distribution(vcf_list, surveyed_list, region_list)
}
\arguments{
\item{vcf_list}{A list with vcf Granges objects}

\item{surveyed_list}{A list with Granges of regions of the genome that have been surveyed (e.g. determined using GATK CallableLoci)}

\item{region_list}{List with GRanges objects containing locations of genomic regions}
}
\value{
A Data.frame containing the number observed and number of expected mutations in each genomic region.
}
\description{
Function finds the number of mutations that reside in genomic region and takes surveyed area of genome into account
}
\examples{
vcf_files = list.files(system.file("extdata",
                                   package="MutationalPatterns"),
                                   pattern = ".vcf",
                                   full.names = TRUE)
sample_names = c("colon1", "colon2", "colon3",
                 "intestine1", "intestine2", "intestine3",
                 "liver1", "liver2", "liver3")
vcfs = read_vcf(vcf_files, sample_names, genome = "hg19")
vcfs = lapply(vcfs, function(x) rename_chrom(x))

# only select autosomal chromosomes, mt dna length is different for vcf and
# ref genome, why??
auto = extractSeqlevelsByGroup(species="Homo_sapiens",
                               style="UCSC",
                               group="auto")
vcfs = lapply(vcfs, function(x) keepSeqlevels(x, auto))

library(biomaRt)
mart="ensemble"
regulation_segmentation = useEnsembl(biomart="regulation",
                                     dataset="hsapiens_segmentation_feature",
                                     GRCh = 37)
regulation_regulatory = useEnsembl(biomart="regulation",
                                   dataset="hsapiens_regulatory_feature",
                                   GRCh = 37)
regulation_annotated = useEnsembl(biomart="regulation",
                                  dataset="hsapiens_annotated_feature",
                                  GRCh = 37)
CTCF = getBM(attributes = c('chromosome_name',
                            'chromosome_start',
                            'chromosome_end',
                            'feature_type_name',
                            'cell_type_name'),
             filters = "regulatory_feature_type_name", 
             values = "CTCF Binding Site", 
             mart = regulation_regulatory)
# Make a GRanges object.
CTCF_g = reduce(GRanges(CTCF$chromosome_name,
                IRanges(CTCF$chromosome_start,
                CTCF$chromosome_end)))

# Get the filename with surveyed/callable regions
surveyed_file = list.files(system.file("extdata",
                                       package="MutationalPatterns"),
                           pattern = ".bed",
                           full.names = TRUE)
# Read the file as a GRanges object.
surveyed_list = bed_to_granges(surveyed_file, "surveyed_all")

# For this example we use the same surveyed file for each sample.
surveyed_list= rep(surveyed_list, 9)

# Calculate the number of observed and expected number of mutations in
# each genomic regions for each sample.
distr = genomic_distribution(vcfs, surveyed_list, regions)

}
\seealso{
\code{\link{bed_to_granges}}
}

