% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_to_signatures_strict.R
\name{get_cos_sim_ori_vs_rec}
\alias{get_cos_sim_ori_vs_rec}
\title{Get the cosine similarity between a reconstructed mutation matrix and the original}
\usage{
get_cos_sim_ori_vs_rec(mut_mat, fit_res)
}
\arguments{
\item{fit_res}{Named list with signature contributions and reconstructed
mutation matrix}

\item{mut_matrix}{mutation count matrix (dimensions: x mutation types
X n samples)}
}
\value{
Cosine similarity
}
\description{
Get the cosine similarity between a reconstructed mutation matrix and the original
}
