% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mut_context.R
\name{get_mut_context}
\alias{get_mut_context}
\title{Retrieve context of base substitutions}
\usage{
get_mut_context(vcf)
}
\arguments{
\item{vcf}{A CollapsedVCF object}

\item{ref_genome}{}
}
\value{
trinucleotides
}
\description{
A function to extract the bases 3' upstream and 5' downstream of each position in vcf
}

