% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mut_type.R
\name{get_mut_type}
\alias{get_mut_type}
\title{Get variants with mut_type from GRanges}
\usage{
get_mut_type(grl, type = c("snv", "indel", "dbs", "mbs"))
}
\arguments{
\item{type}{The type of variant that will be returned.}

\item{gr}{GRanges/GRangesList}
}
\value{
GRanges/GrangesList of the desired mutation type.
}
\description{
Get the variants of a certain mutation type from a GRanges or GRangesList object.
All other variants will be filtered out.
It is assumed that DBS/MBSs are called as separate SNVs.
They are merged into single variants.
The type of variant can be chosen with type.
}
\examples{
## Get a GRanges list object.
## See 'read_vcfs_as_granges' for more info how to do this.
grl <- readRDS(system.file("states/blood_grl.rds",
                package="MutationalPatterns"))

## Get a specific mutation type.
grl_snv = get_mut_type(grl, "snv")
grl_indel = get_mut_type(grl, "indel")
grl_dbs = get_mut_type(grl, "dbs")
grl_mbs = get_mut_type(grl, "mbs")

}
\seealso{
\code{\link{get_mut_type_gr}}, \code{\link{read_vcfs_as_granges}}
}
