% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_muts.R
\name{get_muts}
\alias{get_muts}
\title{Retrieve base substitutions from vcf}
\usage{
get_muts(vcf)
}
\arguments{
\item{vcf}{A CollapsedVCF object}
}
\value{
Character vector with base substitutions
}
\description{
A function to extract base substitutions of each position in vcf
}

