% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_muts.R
\name{get_muts}
\alias{get_muts}
\title{Retrieve base substitutions from vcf}
\usage{
get_muts(vcf)
}
\arguments{
\item{vcf}{A CollapsedVCF object}
}
\value{
Character vector with base substitutions
}
\description{
A function to extract base substitutions of each position in vcf
}
\examples{
## See the 'vcf_to_granges()' example for how we obtained the following data:
vcfs <- readRDS(system.file("states/vcf_to_granges_output.R",
                package="MutationalPatterns"))

muts = get_muts(vcfs[[1]])

}
\seealso{
\code{\link{vcf_to_granges}}
}

