% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lesion_segretation.R
\name{get_strandedness_gr}
\alias{get_strandedness_gr}
\title{Determine the strands of a GRanges object}
\usage{
get_strandedness_gr(gr)
}
\arguments{
\item{gr}{A GRanges object}
}
\value{
A GRanges object where the strands have been set.
}
\description{
Determine the strands of a GRanges object
}
\seealso{
Other Lesion_segregation: 
\code{\link{calculate_lesion_segregation_gr}()},
\code{\link{calculate_lesion_segregation}()},
\code{\link{calculate_strand_switches}()},
\code{\link{calculate_strand_switch}()},
\code{\link{get_strandedness_tb}()},
\code{\link{lesion_get_x_axis_breaks}()},
\code{\link{plot_lesion_segregation_gg}()},
\code{\link{plot_lesion_segregation}()}
}
\concept{Lesion_segregation}
