% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_lesion_segretation.R
\name{get_strandedness_tb}
\alias{get_strandedness_tb}
\title{Convert a GRanges object with strand info to a tibble}
\usage{
get_strandedness_tb(gr)
}
\arguments{
\item{gr}{A GRanges object where the strands have been set.}
}
\value{
A tibble with strand information
}
\description{
Convert a GRanges object with strand info to a tibble
}
