% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_type_context.R
\name{get_type_context}
\alias{get_type_context}
\title{Retrieve context of base substitution types}
\usage{
get_type_context(vcf, ref_genome)
}
\arguments{
\item{vcf}{A CollapsedVCF object}

\item{ref_genome}{Reference genome}
}
\value{
Mutation types and context character vectors in a named list
}
\description{
A function to extract the bases 3' upstream and 5' downstream of the base substitution types
}

