% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_96_occurrences.R
\name{mut_96_occurrences}
\alias{mut_96_occurrences}
\title{Count 96 trinucleotide mutation occurrences}
\usage{
mut_96_occurrences(type_context)
}
\arguments{
\item{type_context}{result from type_context function}
}
\value{
vector with 96 trinucleotide mutation occurrences
}
\description{
Count 96 trinucleotide mutation occurrences
}

