% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_matrix_stranded.R
\name{mut_matrix_stranded}
\alias{mut_matrix_stranded}
\title{Make mutation count matrix of 96 trinucleotides with transcriptional strand information}
\usage{
mut_matrix_stranded(vcf_list, ref_genome, genes)
}
\arguments{
\item{vcf_list}{List of collapsed vcf objects}

\item{ref_genome}{BSGenome reference genome object}

\item{genes}{Granges with definition of gene bodies, should include strand information}
}
\value{
192 mutation count matrix (96 * 2 strands)
}
\description{
Make mutation count matrix for 96 trinucleotides for both transcribed and untranscribed strand of gene bodies. 
Mutations outside gene bodies are not counted.
}

