% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mut_matrix_stranded.R
\name{mut_matrix_stranded}
\alias{mut_matrix_stranded}
\title{Make mutation count matrix of 96 trinucleotides with transcriptional
strand information}
\usage{
mut_matrix_stranded(vcf_list, ref_genome, genes)
}
\arguments{
\item{vcf_list}{List of collapsed vcf objects}

\item{ref_genome}{BSGenome reference genome object}

\item{genes}{GRanges object with definition of gene bodies, including
strand information}
}
\value{
192 mutation count matrix (96 * 2 strands)
}
\description{
Make a mutation count matrix for 96 trinucleotides, for both the
transcribed and untranscribed strand of gene bodies.
}
\details{
Mutations outside gene bodies are not counted.
}
\examples{
## See the 'read_vcfs_as_granges()' example for how we obtained the
## following data:
vcfs <- readRDS(system.file("states/read_vcfs_as_granges_output.rds",
                package="MutationalPatterns"))

## Exclude mitochondrial and allosomal chromosomes.
autosomal <- extractSeqlevelsByGroup(species="Homo_sapiens",
                                    style="UCSC",
                                    group="auto")

vcfs = lapply(vcfs, function(x) keepSeqlevels(x, autosomal))

## Load the corresponding reference genome.
ref_genome = "BSgenome.Hsapiens.UCSC.hg19"
library(ref_genome, character.only = TRUE)

## You can obtain the known genes from the UCSC hg19 dataset using
## Bioconductor:
# source("https://bioconductor.org/biocLite.R")
# biocLite("TxDb.Hsapiens.UCSC.hg19.knownGene")
# library("TxDb.Hsapiens.UCSC.hg19.knownGene")

## For this example, we preloaded the data for you:
genes_hg19 <- readRDS(system.file("states/genes_hg19.rds",
                        package="MutationalPatterns"))

mut_mat_s = mut_matrix_stranded(vcfs, ref_genome, genes_hg19)

}
\seealso{
\code{\link{read_vcfs_as_granges}},
\code{link{mut_matrix}}
}

