\name{MutationalPatterns-defunct}
\alias{mutation_context}
\alias{mutation_types}
\alias{strand_from_vcf}
\alias{explained_by_signatures}
\title{Defunct functions in package \sQuote{MutationalPattern}}
\description{These functions are defunct and no longer available.}

\details{
  Defunct functions are: 
  \code{mutation_context}, 
  \code{mutation_types}, 
  \code{strand_from_vcf}, 
  \code{explained_by_signatures}
}