\name{MutationalPatterns}
\alias{MutationalPatterns-package}
\alias{MutationalPatterns}
\docType{package}
\title{Extracting and visualizing mutational patterns in SNV data}
\description{This package provides a comprehensive set of flexible
  functions for easy finding and plotting of mutational patterns in
  Single Nucleotide Variant (SNV) data.}
\details{
  \tabular{ll}{
    Package: \tab MutationalPatterns\cr
    Type: \tab Package\cr
    Version: \tab 0.1\cr
    Date: \tab 2015-02-09\cr
    License: \tab MIT\cr
  }
  ~~ An overview of how to use the package, including the most important functions ~~
}
\author{Francis Blokzijl <F.Blokzijl@umcutrecht.nl>}
\references{}
\keyword{ package }
\seealso{}
\examples{}
