\name{MutationalPatterns-package}
\alias{MutationalPatterns-package}
\alias{MutationalPatterns}
\docType{package}
\title{
MutationalPatterns: an integrative R package for studying patterns in base substitution catalogues
}
\description{
This package evaluates a wide range of mutational patterns, including: base substitution types, sequence context, mutational signatures, genomic distribution, association with genomic features and transcriptional strand bias.
}
\details{
\tabular{ll}{
Package: MutationalPatterns
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-09\cr
License: MIT + file LICENSE
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Francis Blokzijl, Roel Janssen, Ruben van Boxtel, Edwin Cuppen

Maintainer: Francis Blokzijl, UMC Utrecht <f.blokzijl@umcutrecht.nl>

}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
\url{https://github.com/CuppenResearch/MutationalPatterns}
}
\examples{
~~ simple examples of the most important functions ~~
}
s