% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_192_profile.R
\name{plot_192_profile}
\alias{plot_192_profile}
\title{Plot 192 trinucleotide profile
 
Plot relative contribution of 192 trinucleotides}
\usage{
plot_192_profile(mut_matrix, colors, ymax = 0.15)
}
\arguments{
\item{mut_matrix}{192 trinucleotide profile matrix}

\item{colors}{6 value color vector}

\item{ymax}{Y axis maximum value, default = 0.015}
}
\value{
192 trinucleotide profile plot
}
\description{
Plot 192 trinucleotide profile
 
Plot relative contribution of 192 trinucleotides
}

