% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_96_profile.R
\name{plot_96_profile}
\alias{plot_96_profile}
\title{Plot 96 trinucleotide profile
 
Plot relative contribution of 96 trinucleotides}
\usage{
plot_96_profile(mut_matrix, colors = c("#DBD7C8", "#B2D39C", "#71C1BA",
  "#2DAFCE", "#2476B2", "#737E93"))
}
\arguments{
\item{mut_matrix}{96 trinucleotide profile matrix}
}
\value{
96 trinucleotide profile plot
}
\description{
Plot 96 trinucleotide profile
 
Plot relative contribution of 96 trinucleotides
}

