% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_96_profile.R
\name{plot_96_profile}
\alias{plot_96_profile}
\title{Plot 96 trinucleotide profile
 
Plot relative contribution of 96 trinucleotides}
\usage{
plot_96_profile(mut_matrix, colors = spectrum.colors6, ymax = 0.15)
}
\arguments{
\item{mut_matrix}{96 trinucleotide profile matrix}

\item{colors}{6 value color vector}

\item{ymax}{Y axis maximum value, default = 0.015}
}
\value{
96 trinucleotide profile plot
}
\description{
Plot 96 trinucleotide profile
 
Plot relative contribution of 96 trinucleotides
}

