% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_96_profile.R
\name{plot_96_profile}
\alias{plot_96_profile}
\title{Plot 96 trinucleotide profile}
\usage{
plot_96_profile(mut_matrix, colors, ymax = 0.2, condensed = F)
}
\arguments{
\item{mut_matrix}{96 trinucleotide profile matrix}

\item{colors}{6 value color vector}

\item{ymax}{Y axis maximum value, default = 0.2}

\item{condensed}{More condensed plotting format. Default = F.}
}
\value{
96 trinucleotide profile plot
}
\description{
Plot relative contribution of 96 trinucleotides
}
\examples{
## See the 'mut_matrix_stranded()' example for how we obtained the
## mutation matrix with transcriptional strand information:
mut_mat <- readRDS(system.file("states/mut_mat_data.rds",
                                package="MutationalPatterns"))

## Plot the 96-profile of three samples
plot_96_profile(mut_mat[,c(1,4,7)])

}
\seealso{
\code{\link{mut_matrix}}
}
