% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_profiles.R
\name{plot_compare_profiles}
\alias{plot_compare_profiles}
\title{Compare two 96 mutation profiles}
\usage{
plot_compare_profiles(profile1, profile2, profile_names = c("profile 1",
  "profile 2"), profile_ymax = 0.15, diff_ylim = c(-0.02, 0.02),
  colors = c("#DBD7C8", "#B2D39C", "#71C1BA", "#2DAFCE", "#2476B2",
  "#737E93"))
}
\value{
plot
}
\description{
Compare two 96 mutation profiles
}

