% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contribution.R
\name{plot_contribution}
\alias{plot_contribution}
\title{Plot signature contribution barplot}
\usage{
plot_contribution(contribution, signatures, mut_type, index = c(),
  coord_flip = FALSE, mode = "relative", method = "split",
  palette = c())
}
\arguments{
\item{contribution}{List of signature contribution matrices}

\item{signatures}{List of signature matrices}

\item{index}{(Optional) Sample subset parameter}

\item{coord_flip}{(Optional) Flip X and Y coordinates, default = FALSE}

\item{mode}{(Optional) Character stating to "relative" or "absolute" contribution of
mutations. Also possible to plot "both".\cr 
Default = "relative"}

\item{method}{(Optional) Character stating how to use the data. 
\itemize{
  \item{"split":} { Each mutation type has seperate count matrix}
  \item{"combine":} { Combined count matrix of all mutation types}
}   
Default is "split"}

\item{palette}{(Optional) A list of color palette like c("#FF0000", "#00FF00", "9999CC")
for each mutation type that will be used as colors in the plot.\cr
By default, ggplot2's colors are used to generate a palette.}

\item{type}{(Optional) A character vector stating which type of mutation is to be extracted: 
'snv', 'dbs' and/or 'indel'. All mutation types can also be chosen by 'type = all'.\cr
Default is 'snv'}
}
\value{
Stacked barplot with contribution of each signature for each sample
}
\description{
Plot contribution of signatures
}
\examples{
## See the 'mut_matrix()' example for how we obtained the following
## mutation matrix.
mut_mat <- readRDS(system.file("states/mut_mat_data.rds",
                                package="MutationalPatterns"))

## Extracting signatures can be computationally intensive, so
## we use pre-computed data generated with the following command:
# nmf_res <- extract_signatures(mut_mat, rank = 2)

nmf_res <- readRDS(system.file("states/nmf_res_data.rds",
                    package="MutationalPatterns"))

## Optionally set column and row names.
colnames(nmf_res$signatures) = c("Signature A", "Signature B")
rownames(nmf_res$contribution) = c("Signature A", "Signature B")

## The following are examples of contribution plots.
plot_contribution(nmf_res$contribution,
                    nmf_res$signature,
                    mode = "relative")

plot_contribution(nmf_res$contribution,
                    nmf_res$signature,
                    mode = "absolute")

plot_contribution(nmf_res$contribution,
                    nmf_res$signature,
                    mode = "absolute",
                    index = c(1,2))

plot_contribution(nmf_res$contribution,
                    nmf_res$signature,
                    mode = "absolute",
                    coord_flip = TRUE)

}
\seealso{
\code{\link{extract_signatures}},
\code{\link{mut_matrix}}
}
