% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contribution.R
\name{plot_contribution}
\alias{plot_contribution}
\title{Plot signature contribution}
\usage{
plot_contribution(contribution, signatures, index = c(), coord_flip = F,
  mode = "relative")
}
\arguments{
\item{contribution}{Signature contribution matrix}

\item{signatures}{Signature matrix}

\item{index}{optional sample subset parameter}

\item{coord_flip}{Flip X and Y coordinates, default = F}

\item{mode}{relative or absolute number of mutations, default = "relative"}
}
\description{
Plot contribution of signatures
}

