% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contribution.R
\name{plot_contribution}
\alias{plot_contribution}
\title{Plot signature contribution}
\usage{
plot_contribution(nmf_res)
}
\arguments{
\item{nmf_res}{NMF result}

\item{index}{optional subset parameter}
}
\description{
Plot contribution of signatures including unexplained fraction
}

