% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contribution.R
\name{plot_contribution}
\alias{plot_contribution}
\title{Plot signature contribution}
\usage{
plot_contribution(contribution, signatures, index = c(), coord_flip = FALSE,
  mode = "relative")
}
\arguments{
\item{contribution}{Signature contribution matrix}

\item{signatures}{Signature matrix}

\item{index}{optional sample subset parameter}

\item{coord_flip}{Flip X and Y coordinates, default = FALSE}

\item{mode}{"relative" or "absolute"; to plot the relative contribution or absolute number of mutations, default = "relative"}
}
\value{
Stacked barplot with contribution of each signatures for each sample
}
\description{
Plot contribution of signatures
}

