% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contribution.R
\name{plot_contribution}
\alias{plot_contribution}
\title{Plot signature contribution}
\usage{
plot_contribution(contribution, index = c(), coord_flip = F)
}
\arguments{
\item{index}{optional sample subset parameter}

\item{coord_flip}{Flip X and Y coordinates, default = F}

\item{nmf_res}{NMF result}
}
\description{
Plot contribution of signatures including unexplained fraction
}

