% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contribution_heatmap.R
\name{plot_contribution_heatmap}
\alias{plot_contribution_heatmap}
\title{Plot signature contribution heatmap}
\usage{

}
\arguments{
\item{contribution}{Signature contribution matrix}

\item{sig_order}{Character vector with the desired order of the signature names for plotting. Optional.}

\item{cluster_samples}{Hierarchically cluster samples based on eucledian distance. Default = T.}
}
\value{
Heatmap with relative contribution of each signature for each sample
}
\description{
Plot relative contribution of signatures in heatmap
}
\examples{

}
\seealso{
\code{\link{extract_signatures}},
\code{\link{mut_matrix}}
}
