% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cosine_heatmap.R
\name{plot_cosine_heatmap}
\alias{plot_cosine_heatmap}
\title{Plot cosine similarity heatmap}
\usage{
plot_cosine_heatmap(explained_matrix, sig_order, cluster_samples = T, method = "complete", plot_values = T)
}
\arguments{
\item{sig_order}{Character vector with the desired order of the signature names for plotting. Optional.}

\item{cluster_samples}{Hierarchically cluster samples based on eucledian distance. Default = T.}

\item{method}{The agglomeration method to be used for hierarchical clustering. This should be one of 
"ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) 
or "centroid" (= UPGMC). Default = "complete".}

\item{plot_values}{Plot cosine similarity values in heatmap. Default = F.}

\item{explained}{Matrix with pairwise cosine similarities (dimensions: n samples X n signatures). Result from \code{\link{explained_by_signatures}}}
}
\value{
Heatmap with cosine similarity of each signature with each 96 mutational profile
}
\description{
Plot cosine similarity between mutational profiles and signatures in a heatmap.
}
\details{
The cosine similarity is a value between 0 (distinct) and 1 (identical) and indicates how much of the 96 
mutation profile can be explained by an individual signature. Similar mutational signatures will be equally 
good at explaining a mutational profile. For this reason it is recommended to cluster the signatures based 
on cosine similarity with \code{\link{cluster_signatures}} and plot the signatures in this order, see example.
}
\examples{

## You can download the signatures from the pan-cancer study by
## Alexandrov et al:
# http://cancer.sanger.ac.uk/cancergenome/assets/signatures_probabilities.txt

## We copied the file into our package for your convenience.
filename <- system.file("extdata/signatures_probabilities.txt",
                        package="MutationalPatterns")

cancer_signatures <- read.table(filename, sep = "\\t", header = TRUE)

## Reorder the columns to make the order of the trinucleotide changes compatible.
cancer_signatures <- cancer_signatures[order(cancer_signatures[,1]),]

## Include only the signatures in the matrix.
cancer_signatures <- as.matrix(cancer_signatures[,4:33])

## See the 'mut_matrix()' example for how we obtained the mutation matrix:
mut_mat <- readRDS(system.file("states/mut_mat_data.rds",
                    package="MutationalPatterns"))

## Calculate the cosine similarity between each signature and each 96 mutational profile
explained = explained_by_signatures(mut_mat, cancer_signatures)

## Cluster signatures based on cosine similarity 
sig_hclust = cluster_signatures(cancer_signatures)
sig_order = colnames(cancer_signatures)[sig_hclust$order]

## Plot the cosine similarity between each signature and each sample with hierarchical 
## sample clustering and signatures order based on similarity

plot_cosine_heatmap(explained, sig_order, cluster_samples = T, method = "complete")


}
\seealso{
\code{\link{mut_matrix}},
\code{\link{explained_by_signatures}},
\code{\link{cluster_signatures}}
}
