% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichment_depletion.R
\name{plot_enrichment_depletion}
\alias{plot_enrichment_depletion}
\title{Plot enrichment/depletion of mutations in genomic regions}
\usage{
plot_enrichment_depletion(df)
}
\arguments{
\item{df}{Dataframe result from enrichment_depletion_test()}
}
\value{
Plot with two parts. 1: Barplot with no. mutations expected and observed per region. 2: effect size of enrichment/depletion (log2ratio) with results significance test
}
\description{
Plot enrichment/depletion of mutations in genomic regions
}

