% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichment_depletion.R
\name{plot_enrichment_depletion}
\alias{plot_enrichment_depletion}
\title{Plot enrichment depletion barplot}
\usage{
plot_enrichment_depletion(df)
}
\arguments{
\item{df}{Dataframe result from enrichment_depletion_test()}
}
\value{
plot
}
\description{
Plot enrichment depletion barplot
}

