% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indel_plot_contexts.R
\name{plot_indel_contexts}
\alias{plot_indel_contexts}
\title{Plot the indel contexts}
\usage{
plot_indel_contexts(counts, same_y = F, extra_labels = F)
}
\arguments{
\item{counts}{A tibble containing the number of indels per COSMIC context.}

\item{same_y}{A boolean describing whether the same y axis should be used for all samples.}

\item{extra_labels}{A boolean describing whether extra labels should be added. 
These can clarify the plot, but will shift when different plot widths are used.
We recommend saving a plot with a width of 12, when using this argument.}
}
\value{
A ggplot figure.
}
\description{
Plot the indel contexts
}
\details{
Plots the number of indels  COSMIC context per sample.
It takes a tibble with counts as its input. This tibble can be generated by count_indel_contexts
Each sample is plotted in a separate facet.
The same y axis can be used for all samples or a separate y axis can be used.
}
\examples{
## Get The indel counts
## See 'count_indel_contexts()' for more info on how to do this.
indel_counts <- readRDS(system.file("states/blood_indel_counts.rds",
                package="MutationalPatterns"))

## Plot contexts
plot_indel_contexts(indel_counts)

## Use the same y axis for all samples.
plot_indel_contexts(indel_counts, same_y = TRUE)

## Add extra labels to make plot clearer
plot_indel_contexts(indel_counts, extra_labels = T)

}
\seealso{
\code{\link{count_indel_contexts}}, \code{\link{plot_main_indel_contexts}}

Other Indels: 
\code{\link{count_indel_contexts_gr}()},
\code{\link{count_indel_contexts}()},
\code{\link{get_1bp_dels}()},
\code{\link{get_1bp_ins}()},
\code{\link{get_big_dels}()},
\code{\link{get_big_ins}()},
\code{\link{get_extended_sequence}()},
\code{\link{get_indel_context_gr}()},
\code{\link{get_indel_context}()},
\code{\link{plot_main_indel_contexts}()}
}
\concept{Indels}
