% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rainfall.R
\name{plot_rainfall}
\alias{plot_rainfall}
\title{Plot genomic rainfall}
\usage{
plot_rainfall(vcf, chromosomes, type, method = "split", title = "",
  colors, cex = 2.5, cex_text = 3, ylim = 1e+08)
}
\arguments{
\item{vcf}{CollapsedVCF object}

\item{chromosomes}{Vector of chromosome/contig names of the reference
genome to be plotted}

\item{type}{(Optional) A character vector stating which type of mutation is to be extracted: 
'snv', 'dbs' and/or 'indel'. All mutation types can also be chosen by 'type = all'.\cr
Default is 'snv'}

\item{method}{(Optional) Character stating how to plot the results. method = "split" will give 
seperate plots for each mutation type, whereas method = "combine" will give one plot with all
mutation types.\cr
Default is "split"}

\item{title}{(Optional) Plot title}

\item{colors}{(Optional) Named list with 6 value color vector "snv" for snv, 
10 value color vector "dbs" for dbs.
For indels give same number of colors as there are classes. \cr}

\item{cex}{(Optional) Point size, default = 2.5}

\item{cex_text}{(Optional) Text size, default = 3}

\item{ylim}{(Optional) Maximum y value (genomic distance), default = 1e+8}
}
\value{
Rainfall plot
}
\description{
Rainfall plot visualizes the types of mutations and intermutation distance
}
\details{
Rainfall plots can be used to visualize the distribution of mutations
along the genome or a subset of chromosomes. The distance of a mutation
with the mutation prior to it (the intermutation distance) is plotted on
the y-axis on a log scale.

The colour of the points indicates the base substitution type.
Clusters of mutations with lower intermutation distance represent mutation
hotspots.
}
\examples{
## See the 'read_vcfs_as_granges()' example for how we obtained the
## following data:
vcfs <- readRDS(system.file("states/read_vcfs_as_granges_output.rds",
                package="MutationalPatterns"))

# Specify chromosomes of interest.
chromosomes = names(genome(vcfs[[1]])[1:22])

## Do a rainfall plot for all chromosomes:
plot_rainfall(vcfs[[1]],
                title = names(vcfs[1]),
                chromosomes = chromosomes,
                cex = 1)

## Or for a single chromosome (chromosome 1):
plot_rainfall(vcfs[[1]],
                title = names(vcfs[1]),
                chromosomes = chromosomes[1],
                cex = 2)

}
\seealso{
\code{\link{read_vcfs_as_granges}}
}
