% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signature_strand_bias.R
\name{plot_signature_strand_bias}
\alias{plot_signature_strand_bias}
\title{Plot signature strand bias}
\usage{
plot_signature_strand_bias(signatures_strand_bias)
}
\arguments{
\item{signatures_strand_bias}{Signature matrix with 192 features}
}
\value{
Barplot
}
\description{
Plot strand bias per mutation type for each signature
}

