% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_signatures_96.R
\name{plot_signatures_96}
\alias{plot_signatures_96}
\title{Plot 96 trinucleotide signatures}
\usage{
plot_signatures_96(signatures, colors = c("#DBD7C8", "#B2D39C", "#71C1BA",
  "#2DAFCE", "#2476B2", "#737E93"))
}
\arguments{
\item{vcf_files}{Character vector of vcf_files}

\item{rank}{Number of signatures one would like to extract}

\item{outdir}{Specify output directory}
}
\description{
Plot 96 trinucleotide signatures
}

