% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{Plot point mutation spectrum}
\usage{
plot_spectrum(type_occurences, CT = FALSE, by, colors, legend = TRUE)
}
\arguments{
\item{type_occurences}{Type occurences matrix}

\item{CT}{Distinction between C>T at CpG and C>T at other sites, default = FALSE}

\item{by}{Optional grouping variable}

\item{colors}{Optional color vector with 7 values}

\item{legend}{Plot legend, default = TRUE}
}
\value{
Spectrum plot
}
\description{
Plot point mutation spectrum
}

