% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{Plot point mutation spectrum}
\usage{
plot_spectrum(type_occurences, CT = F, by = "all")
}
\arguments{
\item{type_occurences}{Type occurences matrix}

\item{CT}{Optional distinction between C>T at CpG}

\item{by}{Optional grouping variable}
}
\value{
plot Spectrum plot
}
\description{
Plot point mutation spectrum
}

