% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{Plot point mutation spectrum}
\usage{
plot_spectrum(type_occurences, CT = F, by = "all", colors = c("#DBD7C8",
  "#B2D39C", "#B3D9CF", "#71C1BA", "#2DAFCE", "#2476B2", "#737E93"),
  legend = T)
}
\arguments{
\item{type_occurences}{Type occurences matrix}

\item{CT}{Optional distinction between C>T at CpG}

\item{by}{Optional grouping variable}

\item{colors}{Optional color vector with 7 values}
}
\value{
Spectrum plot
}
\description{
Plot point mutation spectrum
}

