% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_strand.R
\name{plot_strand}
\alias{plot_strand}
\title{Plot strand per base substitution type}
\usage{
plot_strand(strand_counts, mode = "relative", colors)
}
\arguments{
\item{strand_counts}{A data.frame, result from strand_occurrences function}

\item{mode}{(Optional) Either "absolute" for absolute number of mutations, 
"relative" for relative contribution or "both" for both.\cr
Default = "relative"}

\item{colors}{(Optional) Named list with 6 value color vector "snv" for snv, 
10 value color vector "dbs" for dbs.
For indels give same number of colors as there are classes.}
}
\value{
Barplot
}
\description{
For each base substitution type and transcriptional strand the total number
of mutations and the relative contribution within a group is returned.
}
\examples{
## See the 'mut_matrix_stranded()' example for how we obtained the
## following mutation matrix.
mut_mat_s <- readRDS(system.file("states/mut_mat_s_data.rds",
                                    package="MutationalPatterns"))

## Load a reference genome.
ref_genome <- "BSgenome.Hsapiens.UCSC.hg19"
library(ref_genome, character.only = TRUE)

tissue <- c("colon", "colon", "colon",
            "intestine", "intestine", "intestine",
            "liver", "liver", "liver")

strand_counts = strand_occurrences(mut_mat_s, by=tissue)

#' ## Plot the strand in relative mode.
strand_plot = plot_strand(strand_counts)

#' ## Or absolute mode.
strand_plot = plot_strand(strand_counts, mode = "absolute")

}
\seealso{
\code{\link{mut_matrix_stranded}},
\code{\link{strand_occurrences}},
\code{\link{plot_strand_bias}}
}
