% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_strand_bias.R
\name{plot_strand_bias}
\alias{plot_strand_bias}
\title{Plot strand bias per base substitution type per group}
\usage{
plot_strand_bias(strand_bias, colors, max_yaxis = FALSE)
}
\arguments{
\item{strand_bias}{A data.frame, result from strand_bias function}

\item{colors}{(Optional) Named list with 6 value color vector "snv" for snv, 
10 value color vector "dbs" for dbs.
For indels give same number of colors as there are classes.}

\item{max_yaxis}{(Optional) Binary option to set y axis to maximum value of all
mutation types. Default is 'max = FALSE'}
}
\value{
Barplot
}
\description{
Plot strand bias per base substitution type per group
}
\examples{
## See the 'mut_matrix_stranded()' example for how we obtained the
## following mutation matrix.
mut_mat_s <- readRDS(system.file("states/mut_mat_s_data.rds",
                                    package="MutationalPatterns"))

## Load a reference genome.
ref_genome <- "BSgenome.Hsapiens.UCSC.hg19"
library(ref_genome, character.only = TRUE)

tissue <- c("colon", "colon", "colon",
            "intestine", "intestine", "intestine",
            "liver", "liver", "liver")

## Perform the strand bias test.
strand_counts = strand_occurrences(mut_mat_s, by=tissue)
strand_bias = strand_bias_test(strand_counts)

## Plot the strand bias.
plot_strand_bias(strand_bias)

}
\seealso{
\code{\link{mut_matrix_stranded}},
\code{\link{strand_occurrences}},
\code{\link{strand_bias_test}}
\code{\link{plot_strand}}
}
