% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_strand_bias.R
\name{plot_strand_bias}
\alias{plot_strand_bias}
\title{Plot strand bias}
\usage{
plot_strand_bias(strand_bias_df, mode = "relative", colors)
}
\arguments{
\item{strand_bias_df}{the data frame containing strand bias data}

\item{mode}{Either "absolute" or "relative"}

\item{colors}{the colors to use in the plot}
}
\value{
a plot of the strand bias data
}
\description{
A function to plot the absolute or relative strand bias
}

