% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_strand_bias.R
\name{plot_strand_bias}
\alias{plot_strand_bias}
\title{Plot strand bias per base substitution type}
\usage{
plot_strand_bias(strand_bias, colors)
}
\arguments{
\item{strand_bias}{data.frame, result from strand_bias function}

\item{colors}{Optional color vector for plotting with 6 values}
}
\value{
Barplot
}
\description{
For each base substitution type and transcriptional strand the total number of mutations
and the relative contribution within a group is returned
}

