% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_strand_bias.R
\name{plot_strand_bias}
\alias{plot_strand_bias}
\title{Plot strand bias per base substitution type}
\usage{
plot_strand_bias(strand_bias_df, mode = "relative", colors)
}
\arguments{
\item{strand_bias_df}{Data.frame, result from strand_bias function}

\item{mode}{Either "absolute" for absolute number of mutations, or "relative" for relative contribution, default = "relative"}

\item{colors}{Optional color vector for plotting with 6 values}
}
\value{
Barplot
}
\description{
For each base substitution type and transcriptional strand the total number of mutations
and the relative contribution within a group is returned
}

