% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall_plot.R
\name{rainfall_plot}
\alias{rainfall_plot}
\title{Genomic rainfall plot}
\usage{
rainfall_plot(vcf, ref_genome, chromosomes, title = "", color = c("#1B9E77",
  "#D95F02", "#7570B3", "#E7298A", "#E6AB02", "black"), cex = 2.5,
  cex_text = 3, ylim = 1e+08)
}
\arguments{
\item{vcf}{CollapsedVCF object}

\item{ref_genome}{BSgenome reference genome object}

\item{title}{Optional plot title}

\item{color}{Vector of 6 colors used for plotting}

\item{cex}{Point size}

\item{cex_text}{Text size}

\item{ylim}{Maximum y value (genomic distance)}

\item{chrom}{Vector of chromosome/contig names of the reference genome to be plotted}
}
\value{
Rainfall plot
}
\description{
Rainfall plot visualizes the types of mutations and intermutation distance
}

