% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf.R
\name{read_vcf}
\alias{read_vcf}
\title{Read vcf file}
\usage{
read_vcf(vcf_file)
}
\arguments{
\item{vcf_file}{Vcf file to be read}
}
\value{
A CollapsedVCF object
}
\description{
Function reads Variant Call Format (VCF) file into a CollapsedVCF object for hg19 genome
}

