% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_chrom.R
\name{rename_chrom}
\alias{rename_chrom}
\title{Rename chromosome names to standard}
\usage{
rename_chrom(granges, style = "UCSC")
}
\arguments{
\item{granges}{Granges object with chromosome names to rename}

\item{style}{"UCSC", NCBI" or "Ensembl", default = "UCSC"}
}
\value{
granges Granges object with renamed chromosomes according to style
}
\description{
Rename chromosome names to standard
}

