% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_context_dbs.R
\name{set_context_dbs_gr}
\alias{set_context_dbs_gr}
\title{Set DBS context on a GRanges}
\usage{
set_context_dbs_gr(gr)
}
\arguments{
\item{gr}{GRanges}
}
\value{
A version of the GRanges object, with modified REF and ALT columns.
}
\description{
Set the DBS COSMIC context on an GRanges object.
This is done by changing the REF and ALT columns of the GRanges into the COSMIC types.
}
\seealso{
\code{\link{set_context_dbs}}
}
