% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand_bias.R
\name{strand_bias}
\alias{strand_bias}
\title{Calculate strand bias}
\usage{
strand_bias(mut_mat_s, by)
}
\arguments{
\item{mut_mat_s}{The mutation matrix}

\item{by}{The column to group by}
}
\value{
A data frame with strand bias data
}
\description{
A function to calculate strand bias.
}

