% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand_bias_test.R
\name{strand_bias_test}
\alias{strand_bias_test}
\title{Significance test for transcriptional strand asymmetry}
\usage{
strand_bias_test(strand_occurences)
}
\arguments{
\item{strand_occurences}{Dataframe with mutation count per strand, result from strand_occurences()}
}
\value{
Dataframe with poisson test P value for the ratio between the transcribed and untrascribed strand 
per group per base substitution type
}
\description{
Performs a Poisson test for the ratio between mutations on the transcribed and untranscribed strand
}

