% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand_from_vcf.R
\name{strand_from_vcf}
\alias{strand_from_vcf}
\title{Find transcriptional strand of base substitutions in vcf}
\usage{
strand_from_vcf(vcf, genes)
}
\arguments{
\item{vcf}{GRanges containing the VCF object}

\item{genes}{GRanges with gene bodies definitions including strand
information}
}
\value{
Character vector with transcriptional strand information with
length of vcf: "-" for positions outside gene bodies, "U" for
untranscribed/sense/coding strand, "T" for
transcribed/anti-sense/non-coding strand.
}
\description{
For the positions that are within gene bodies it is determined whether
the "C" or "T" base is on the same strand as the gene definition. (Since
by convention we regard base substitutions as C>X or T>X.)
}
\details{
Base substitions on the same strand as the gene definitions are considered
untranscribed, and on the opposite strand of gene bodies as transcribed,
since the gene definitions report the coding or sense strand, which is
untranscribed.

No strand information "-" is returned for base substitutions outside gene
bodies, or base substitutions that overlap with more than one gene body.
}
\examples{
## For this example we need our variants from the VCF samples, and
## a known genes dataset.  See the 'read_vcfs_as_granges()' example
## for how to load the VCF samples.
vcfs <- readRDS(system.file("states/read_vcfs_as_granges_output.rds",
                package="MutationalPatterns"))

# Exclude mitochondrial and allosomal chromosomes.
autosomal = extractSeqlevelsByGroup(species="Homo_sapiens",
                                    style="UCSC",
                                    group="auto")

vcfs = lapply(vcfs, function(x) keepSeqlevels(x, autosomal))

## You can obtain the known genes from the UCSC hg19 dataset using
## Bioconductor:
# source("https://bioconductor.org/biocLite.R")
# biocLite("TxDb.Hsapiens.UCSC.hg19.knownGene")
# library("TxDb.Hsapiens.UCSC.hg19.knownGene")

## For this example, we preloaded the data for you:
genes_hg19 <- readRDS(system.file("states/genes_hg19.rds",
                        package="MutationalPatterns"))

strand_from_vcf(vcfs[[1]], genes_hg19)

}
\seealso{
\code{\link{read_vcfs_as_granges}},
}

